refactor: distribute remaining impl methods to submodules
Moves the last ~3300 lines of methods from mod.rs into existing
submodules: repeat/keymaps/mouse to keys.rs, workspace to buffers.rs,
bracket nav/toggle case/join/scroll/jump list/indent to motions.rs,
window resize to windows.rs, search word to search.rs, multi-cursor
to visual.rs.

mod.rs is now ~3335 lines (types, struct, new(), free functions).
All 4721 tests pass.
