+++ src/core/engine/mod.rs
@@ -808,7 +808,7 @@ struct Change {
 }
 
 /// C preprocessor directive kind for `[#` / `]#` navigation.
-enum PreprocKind {
+pub(crate) enum PreprocKind {
     If,
     ElseElif,
     Endif,
@@ -2861,3295 +2861,6 @@ impl Engine {
 
         engine.plugin_init();
         engine
-    }
-
-    // =======================================================================
-    // Repeat command (.)
-    // =======================================================================
-
-    fn repeat_last_change(&mut self, repeat_count: usize, changed: &mut bool) {
-        let change = match &self.last_change {
-            Some(c) => c.clone(),
-            None => return, // No change to repeat
-        };
-
-        let final_count = if repeat_count > 1 {
-            repeat_count
-        } else {
-            change.count
-        };
-
-        match change.op {
-            ChangeOp::Insert => {
-                // Repeat insert: insert the same text at current position
-                self.start_undo_group();
-                let line = self.view().cursor.line;
-                let col = self.view().cursor.col;
-                let char_idx = self.buffer().line_to_char(line) + col;
-
-                // Insert the text final_count times
-                let repeated_text = change.text.repeat(final_count);
-                self.insert_with_undo(char_idx, &repeated_text);
-
-                // Update cursor position based on inserted text
-                let newlines = repeated_text.matches('\n').count();
-                if newlines > 0 {
-                    self.view_mut().cursor.line += newlines;
-                    // Find column after last newline
-                    if let Some(last_nl) = repeated_text.rfind('\n') {
-                        self.view_mut().cursor.col = repeated_text[last_nl + 1..].chars().count();
-                    }
-                } else {
-                    self.view_mut().cursor.col += repeated_text.chars().count();
-                }
-                self.finish_undo_group();
-                *changed = true;
-            }
-            ChangeOp::Delete => {
-                // Repeat delete with motion
-                if let Some(motion) = &change.motion {
-                    for _ in 0..final_count {
-                        self.start_undo_group();
-                        match motion {
-                            Motion::Right => {
-                                // Delete character(s) at cursor (like x)
-                                let line = self.view().cursor.line;
-                                let col = self.view().cursor.col;
-                                let char_idx = self.buffer().line_to_char(line) + col;
-                                let line_end = self.buffer().line_to_char(line)
-                                    + self.buffer().line_len_chars(line);
-                                let available = line_end - char_idx;
-                                let to_delete = change.count.min(available);
-
-                                if to_delete > 0 && char_idx < self.buffer().len_chars() {
-                                    let deleted_chars: String = self
-                                        .buffer()
-                                        .content
-                                        .slice(char_idx..char_idx + to_delete)
-                                        .chars()
-                                        .collect();
-                                    let reg = self.active_register();
-                                    self.set_register(reg, deleted_chars, false);
-                                    self.clear_selected_register();
-                                    self.delete_with_undo(char_idx, char_idx + to_delete);
-                                    self.clamp_cursor_col();
-                                    *changed = true;
-                                }
-                            }
-                            Motion::DeleteLine => {
-                                // Repeat dd
-                                self.delete_lines(change.count, changed);
-                            }
-                            _ => {}
-                        }
-                        self.finish_undo_group();
-                    }
-                }
-            }
-            ChangeOp::Change => {
-                // Repeat c{motion}: delete the motion range, then insert the text.
-                if let Some(motion) = &change.motion {
-                    for _ in 0..final_count {
-                        let motion_char = match motion {
-                            Motion::WordForward => 'w',
-                            Motion::WordEnd => 'e',
-                            Motion::WordBackward => 'b',
-                            _ => continue,
-                        };
-                        // Reuse the same code path as the original cw/ce/cb:
-                        // apply_operator_with_motion deletes the range and enters
-                        // insert mode.  We then immediately insert the recorded
-                        // text and return to normal mode instead.
-                        let start_cursor = self.view().cursor;
-                        let start_pos =
-                            self.buffer().line_to_char(start_cursor.line) + start_cursor.col;
-                        for _ in 0..change.count {
-                            match motion_char {
-                                'w' => self.move_word_forward(),
-                                'b' => self.move_word_backward(),
-                                'e' => self.move_word_end(),
-                                _ => {}
-                            }
-                        }
-                        let end_cursor = self.view().cursor;
-                        let end_pos = self.buffer().line_to_char(end_cursor.line) + end_cursor.col;
-                        self.view_mut().cursor = start_cursor;
-
-                        let delete_end = if motion_char == 'e' {
-                            (end_pos + 1).min(self.buffer().len_chars())
-                        } else {
-                            end_pos
-                        };
-                        if start_pos < delete_end {
-                            self.start_undo_group();
-                            self.delete_with_undo(start_pos, delete_end);
-                            if !change.text.is_empty() {
-                                self.insert_with_undo(start_pos, &change.text);
-                                let inserted_chars = change.text.chars().count();
-                                let newlines = change.text.matches('\n').count();
-                                if newlines > 0 {
-                                    self.view_mut().cursor.line += newlines;
-                                    if let Some(last_nl) = change.text.rfind('\n') {
-                                        self.view_mut().cursor.col =
-                                            change.text[last_nl + 1..].chars().count();
-                                    }
-                                } else {
-                                    self.view_mut().cursor.col += inserted_chars;
-                                }
-                            }
-                            self.clamp_cursor_col();
-                            self.finish_undo_group();
-                            *changed = true;
-                        }
-                    }
-                }
-            }
-            ChangeOp::Substitute => {
-                // Repeat s command
-                for _ in 0..final_count {
-                    let line = self.view().cursor.line;
-                    let col = self.view().cursor.col;
-                    let max_col = self.get_max_cursor_col(line);
-                    if max_col > 0 || self.buffer().line_len_chars(line) > 0 {
-                        let char_idx = self.buffer().line_to_char(line) + col;
-                        let line_end =
-                            self.buffer().line_to_char(line) + self.buffer().line_len_chars(line);
-                        let available = line_end - char_idx;
-                        let to_delete = change.count.min(available);
-
-                        self.start_undo_group();
-                        if to_delete > 0 && char_idx < self.buffer().len_chars() {
-                            self.delete_with_undo(char_idx, char_idx + to_delete);
-                            *changed = true;
-                        }
-
-                        // Insert the recorded text
-                        if !change.text.is_empty() {
-                            self.insert_with_undo(char_idx, &change.text);
-                            *changed = true;
-                        }
-                        self.finish_undo_group();
-                    }
-                }
-            }
-            ChangeOp::SubstituteLine | ChangeOp::DeleteToEnd | ChangeOp::ChangeToEnd => {
-                // Handle other operations
-            }
-            ChangeOp::Replace => {
-                // Repeat r command
-                if let Some(replacement_char) = change.text.chars().next() {
-                    for _ in 0..final_count {
-                        self.start_undo_group();
-                        self.replace_chars(replacement_char, change.count, changed);
-                        self.finish_undo_group();
-                    }
-                }
-            }
-            ChangeOp::ToggleCase => {
-                // Repeat ~ command
-                for _ in 0..final_count {
-                    self.toggle_case_at_cursor(change.count, changed);
-                }
-            }
-            ChangeOp::Join => {
-                // Repeat J command
-                for _ in 0..final_count {
-                    self.join_lines(change.count, changed);
-                }
-            }
-            ChangeOp::Indent => {
-                // Repeat >> command
-                let line = self.view().cursor.line;
-                for _ in 0..final_count {
-                    self.indent_lines(line, change.count, changed);
-                }
-            }
-            ChangeOp::Dedent => {
-                // Repeat << command
-                let line = self.view().cursor.line;
-                for _ in 0..final_count {
-                    self.dedent_lines(line, change.count, changed);
-                }
-            }
-        }
-    }
-
-    /// Available commands for auto-completion
-    fn available_commands() -> &'static [&'static str] {
-        &[
-            // File operations
-            "w",
-            "q",
-            "q!",
-            "wq",
-            "wq!",
-            "wa",
-            "wqa",
-            "qa",
-            "qa!",
-            "e ",
-            "e!",
-            "enew",
-            // Buffers
-            "bn",
-            "bp",
-            "bd",
-            "b#",
-            "ls",
-            "buffers",
-            "files",
-            // Splits & tabs
-            "split",
-            "vsplit",
-            "close",
-            "only",
-            "new",
-            "wincmd ",
-            "tabnew",
-            "tabnext",
-            "tabprev",
-            "tabclose",
-            "tabmove",
-            // Search & replace
-            "s/",
-            "%s/",
-            "noh",
-            "nohlsearch",
-            // Settings & config
-            "set ",
-            "config reload",
-            "Settings",
-            "Keymaps",
-            "Keybindings",
-            "Keybindings ",
-            "colorscheme ",
-            // Editor groups
-            "EditorGroupSplit",
-            "EditorGroupSplitDown",
-            "EditorGroupClose",
-            "EditorGroupFocus",
-            "EditorGroupMoveTab",
-            "egsp",
-            "egspd",
-            "egc",
-            "egf",
-            "egmt",
-            // Netrw / file browser
-            "Explore",
-            "Ex",
-            "Sexplore",
-            "Sex",
-            "Vexplore",
-            "Vex",
-            // Git
-            "Gdiff",
-            "Gd",
-            "Gdiffsplit",
-            "Gds",
-            "Gstatus",
-            "Gs",
-            "Gadd",
-            "Ga",
-            "Gcommit",
-            "Gc",
-            "Gpush",
-            "Gp",
-            "Gblame",
-            "Gb",
-            "Ghs",
-            "Ghunk",
-            "Gpull",
-            "Gfetch",
-            "Gswitch",
-            "GSwitch",
-            "Gsw",
-            "Gbranch",
-            "GBranch",
-            "GWorktreeAdd",
-            "GWorktreeRemove",
-            "DiffPeek",
-            "DiffNext",
-            "DiffPrev",
-            "DiffToggleContext",
-            // LSP
-            "LspInfo",
-            "LspRestart",
-            "LspStop",
-            "LspInstall",
-            "Lformat",
-            "Rename",
-            "def",
-            "refs",
-            "hover",
-            "LspImpl",
-            "LspTypedef",
-            "CodeAction",
-            // Navigation
-            "nextdiag",
-            "prevdiag",
-            "nexthunk",
-            "prevhunk",
-            "fuzzy",
-            "sidebar",
-            "palette",
-            // DAP / Debug
-            "DapInfo",
-            "DapInstall",
-            "DapCondition",
-            "DapHitCondition",
-            "DapLogMessage",
-            "DapWatch",
-            "DapBottomPanel",
-            "DapEval",
-            "DapExpand",
-            "debug",
-            "continue",
-            "pause",
-            "stop",
-            "restart",
-            "stepover",
-            "stepin",
-            "stepout",
-            "brkpt",
-            // Extensions
-            "ExtInstall",
-            "ExtList",
-            "ExtEnable",
-            "ExtDisable",
-            "ExtRemove",
-            "ExtRefresh",
-            // AI
-            "AI ",
-            "AiClear",
-            // Markdown
-            "MarkdownPreview",
-            "MdPreview",
-            // Display / info
-            "registers",
-            "display",
-            "marks",
-            "jumps",
-            "changes",
-            "history",
-            "echo ",
-            // Diff
-            "diffthis",
-            "diffoff",
-            "diffsplit",
-            // Misc ex commands
-            "sort",
-            "terminal",
-            "cd ",
-            "make",
-            "copen",
-            "cn",
-            "cp",
-            "cc",
-            "r ",
-            "norm ",
-            "Plugin",
-            "map",
-            "unmap",
-        ]
-    }
-
-    /// All setting names recognized by `:set`.
-    fn setting_names() -> &'static [&'static str] {
-        &[
-            // Boolean options (full names + aliases)
-            "number",
-            "nu",
-            "relativenumber",
-            "rnu",
-            "expandtab",
-            "et",
-            "autoindent",
-            "ai",
-            "incsearch",
-            "is",
-            "lsp",
-            "wrap",
-            "hlsearch",
-            "hls",
-            "ignorecase",
-            "ic",
-            "smartcase",
-            "scs",
-            "cursorline",
-            "cul",
-            "autoread",
-            "ar",
-            "splitbelow",
-            "sb",
-            "splitright",
-            "spr",
-            "ai_completions",
-            "formatonsave",
-            "fos",
-            "showhiddenfiles",
-            "shf",
-            "swapfile",
-            "breadcrumbs",
-            "autohidepanels",
-            // Value options
-            "tabstop",
-            "ts",
-            "shiftwidth",
-            "sw",
-            "scrolloff",
-            "so",
-            "colorcolumn",
-            "cc",
-            "textwidth",
-            "tw",
-            "updatetime",
-            "ut",
-            "mode",
-            "filetype",
-            "ft",
-        ]
-    }
-
-    /// Find completions for partial command, including argument completion.
-    fn complete_command(&self, partial: &str) -> Vec<String> {
-        if partial.is_empty() {
-            return Vec::new();
-        }
-
-        // Check if we're completing an argument (text after a space)
-        if let Some(space_pos) = partial.find(' ') {
-            let cmd_prefix = &partial[..space_pos];
-            let arg_partial = partial[space_pos + 1..].trim_start();
-
-            return match cmd_prefix {
-                "set" => {
-                    // Complete setting names, including "no" prefixed variants
-                    let mut results: Vec<String> = Self::setting_names()
-                        .iter()
-                        .filter(|name| name.starts_with(arg_partial))
-                        .map(|name| format!("set {name}"))
-                        .collect();
-                    // Also offer "no" prefixed boolean disable variants
-                    for name in Self::setting_names() {
-                        let no_name = format!("no{name}");
-                        if no_name.starts_with(arg_partial) && !arg_partial.is_empty() {
-                            results.push(format!("set {no_name}"));
-                        }
-                    }
-                    results.sort();
-                    results.dedup();
-                    results
-                }
-                "Keybindings" | "keybindings" => ["vim", "vscode"]
-                    .iter()
-                    .filter(|m| m.starts_with(arg_partial))
-                    .map(|m| format!("Keybindings {m}"))
-                    .collect(),
-                "colorscheme" => {
-                    // Complete theme names
-                    let mut names = vec![
-                        "onedark".to_string(),
-                        "gruvbox-dark".to_string(),
-                        "tokyo-night".to_string(),
-                        "solarized-dark".to_string(),
-                        "vscode-dark".to_string(),
-                        "vscode-light".to_string(),
-                        "gruvbox".to_string(),
-                        "tokyonight".to_string(),
-                        "solarized".to_string(),
-                    ];
-                    names.extend(list_custom_theme_names());
-                    names.sort();
-                    names.dedup();
-                    names
-                        .into_iter()
-                        .filter(|name| name.starts_with(arg_partial))
-                        .map(|name| format!("colorscheme {name}"))
-                        .collect()
-                }
-                _ => {
-                    // For other commands with trailing space in available_commands,
-                    // fall through to prefix matching
-                    Self::available_commands()
-                        .iter()
-                        .filter(|cmd| cmd.starts_with(partial))
-                        .map(|s| s.to_string())
-                        .collect()
-                }
-            };
-        }
-
-        // First-word completion
-        Self::available_commands()
-            .iter()
-            .filter(|cmd| cmd.starts_with(partial))
-            .map(|s| s.to_string())
-            .collect()
-    }
-
-    /// Find common prefix of strings
-    fn find_common_prefix(strings: &[String]) -> String {
-        if strings.is_empty() {
-            return String::new();
-        }
-
-        let first = &strings[0];
-        let mut common = String::new();
-
-        for (i, ch) in first.chars().enumerate() {
-            if strings.iter().all(|s| s.chars().nth(i) == Some(ch)) {
-                common.push(ch);
-            } else {
-                break;
-            }
-        }
-
-        common
-    }
-
-    // ─── Workspace / Open Folder ──────────────────────────────────────────────
-
-    /// Open a folder as the new working directory.  Clears all buffers/tabs,
-    /// resets the explorer root, and loads any per-project session state.
-    pub fn open_folder(&mut self, path: &Path) {
-        let canonical = path.canonicalize().unwrap_or_else(|_| path.to_path_buf());
-
-        // Save current per-workspace session before switching
-        if let Some(ref root) = self.workspace_root.clone() {
-            self.save_session_for_workspace(root);
-        }
-
-        // Restore user settings baseline before applying any new folder overlay
-        if let Some(base) = self.base_settings.take() {
-            self.settings = *base;
-        }
-
-        // Check if the new folder has a per-folder settings file to apply as overlay
-        let folder_settings_path = canonical.join(".vimcode").join("settings.json");
-        if folder_settings_path.exists() {
-            if let Ok(content) = std::fs::read_to_string(&folder_settings_path) {
-                if let Ok(json) = serde_json::from_str::<serde_json::Value>(&content) {
-                    if let Some(obj) = json.as_object() {
-                        // Save baseline before overlay
-                        self.base_settings = Some(Box::new(self.settings.clone()));
-                        for (key, value) in obj {
-                            let arg = match value {
-                                serde_json::Value::Bool(b) => {
-                                    if *b {
-                                        key.clone()
-                                    } else {
-                                        format!("no{}", key)
-                                    }
-                                }
-                                serde_json::Value::Number(n) => format!("{}={}", key, n),
-                                serde_json::Value::String(s) => format!("{}={}", key, s),
-                                _ => continue,
-                            };
-                            self.settings.parse_set_option(&arg).ok();
-                        }
-                    }
-                }
-            }
-        }
-
-        // Delete swap files for all current buffers before discarding them.
-        self.cleanup_all_swaps();
-
-        // Clear all existing buffers and tabs, reset to single empty window
-        self.buffer_manager = super::buffer_manager::BufferManager::new();
-        let buffer_id = self.buffer_manager.create();
-        let window_id = super::window::WindowId(self.next_window_id);
-        self.next_window_id += 1;
-        let window = super::window::Window::new(window_id, buffer_id);
-        self.windows.clear();
-        self.windows.insert(window_id, window);
-        let tab = super::tab::Tab::new(super::tab::TabId(self.next_tab_id), window_id);
-        self.next_tab_id += 1;
-        self.editor_groups.clear();
-        let gid = GroupId(0);
-        self.editor_groups.insert(gid, EditorGroup::new(tab));
-        self.active_group = gid;
-        self.group_layout = GroupLayout::leaf(gid);
-        self.next_group_id = 1;
-        self.mode = Mode::Normal;
-
-        // Update cwd + workspace root + process working directory
-        self.cwd = canonical.clone();
-        self.workspace_root = Some(canonical.clone());
-        let _ = std::env::set_current_dir(&canonical);
-
-        // Update git branch
-        self.git_branch = git::current_branch(&canonical);
-
-        // Load per-project session (restores open files + positions)
-        let ws_session = SessionState::load_for_workspace(&canonical);
-        // Restore open files from session
-        let open_files: Vec<PathBuf> = ws_session.open_files.clone();
-        let active_file = ws_session.active_file.clone();
-        // Merge relevant session fields
-        self.session.file_positions = ws_session.file_positions;
-        // Add to recent workspaces in global session
-        self.session.add_recent_workspace(&canonical);
-
-        // Re-open session files
-        for fp in &open_files {
-            self.open_file_in_tab(fp);
-        }
-        // Focus the previously active file
-        if let Some(ref af) = active_file {
-            self.open_file_in_tab(af);
-        }
-
-        self.message = format!("Opened folder: {}", canonical.display());
-    }
-
-    /// Parse and load a `.vimcode-workspace` JSON file.
-    pub fn open_workspace(&mut self, ws_path: &Path) {
-        let content = match std::fs::read_to_string(ws_path) {
-            Ok(c) => c,
-            Err(e) => {
-                self.message = format!("Cannot read workspace: {}", e);
-                return;
-            }
-        };
-        let json: serde_json::Value = match serde_json::from_str(&content) {
-            Ok(v) => v,
-            Err(e) => {
-                self.message = format!("Invalid workspace JSON: {}", e);
-                return;
-            }
-        };
-
-        // Resolve folder path relative to workspace file
-        let ws_dir = ws_path.parent().unwrap_or(Path::new("."));
-        let folder_rel = json
-            .get("folders")
-            .and_then(|f| f.as_array())
-            .and_then(|a| a.first())
-            .and_then(|e| e.get("path"))
-            .and_then(|p| p.as_str())
-            .unwrap_or(".");
-        let folder_path = ws_dir.join(folder_rel);
-        self.workspace_file = Some(ws_path.to_path_buf());
-
-        // Apply any settings overrides from workspace
-        if let Some(settings_obj) = json.get("settings").and_then(|s| s.as_object()) {
-            // Save baseline settings before applying workspace overlay (once only)
-            if self.base_settings.is_none() {
-                self.base_settings = Some(Box::new(self.settings.clone()));
-            }
-            for (key, value) in settings_obj {
-                let arg = match value {
-                    serde_json::Value::Bool(b) => {
-                        if *b {
-                            key.clone()
-                        } else {
-                            format!("no{}", key)
-                        }
-                    }
-                    serde_json::Value::Number(n) => format!("{}={}", key, n),
-                    serde_json::Value::String(s) => format!("{}={}", key, s),
-                    _ => continue,
-                };
-                self.settings.parse_set_option(&arg).ok();
-            }
-        }
-
-        self.open_folder(&folder_path);
-        self.message = format!("Workspace loaded: {}", ws_path.display());
-    }
-
-    /// Write a `.vimcode-workspace` file at the given path with the current folder.
-    /// Open or create a workspace in the directory of the currently active file.
-    /// If a `.vimcode-workspace` file already exists there, open it; otherwise
-    /// create one and then open the folder.
-    pub fn open_workspace_from_file(&mut self) {
-        let buf_id = self.active_window().buffer_id;
-        let dir = self
-            .buffer_manager
-            .get(buf_id)
-            .and_then(|bs| bs.file_path.as_ref())
-            .and_then(|fp| fp.parent())
-            .map(|p| p.to_path_buf())
-            .unwrap_or_else(|| self.cwd.clone());
-        let ws_path = dir.join(".vimcode-workspace");
-        if ws_path.exists() {
-            self.open_workspace(&ws_path);
-        } else {
-            self.save_workspace_as(&ws_path);
-            self.open_folder(&dir);
-        }
-    }
-
-    pub fn save_workspace_as(&mut self, ws_path: &Path) {
-        let folder_path = if let Some(parent) = ws_path.parent() {
-            // Make folder path relative to workspace file location
-            let canonical_cwd = self.cwd.canonicalize().unwrap_or_else(|_| self.cwd.clone());
-            let canonical_parent = parent
-                .canonicalize()
-                .unwrap_or_else(|_| parent.to_path_buf());
-            if canonical_cwd == canonical_parent {
-                ".".to_string()
-            } else {
-                canonical_cwd.to_string_lossy().into_owned()
-            }
-        } else {
-            ".".to_string()
-        };
-
-        let ws = serde_json::json!({
-            "version": 1,
-            "folders": [{"path": folder_path}],
-            "settings": {}
-        });
-        match std::fs::write(
-            ws_path,
-            serde_json::to_string_pretty(&ws).unwrap_or_default(),
-        ) {
-            Ok(()) => {
-                self.workspace_file = Some(ws_path.to_path_buf());
-                self.message = format!("Workspace saved: {}", ws_path.display());
-            }
-            Err(e) => {
-                self.message = format!("Cannot save workspace: {}", e);
-            }
-        }
-    }
-
-    /// Save per-workspace session state (open files, cursor positions).
-    pub fn save_session_for_workspace(&self, root: &Path) {
-        let mut ws_session = SessionState::default();
-
-        // Collect open file paths per group (by iterating each group's tabs).
-        let files_for_group = |group: &EditorGroup| -> Vec<PathBuf> {
-            let mut files: Vec<PathBuf> = Vec::new();
-            for tab in &group.tabs {
-                if let Some(window) = self.windows.get(&tab.active_window) {
-                    if let Some(bs) = self.buffer_manager.get(window.buffer_id) {
-                        if let Some(ref fp) = bs.file_path {
-                            if !files.contains(fp) {
-                                files.push(fp.clone());
-                            }
-                        }
-                    }
-                }
-            }
-            files
-        };
-
-        let group_ids = self.group_layout.group_ids();
-        if let Some(gid) = group_ids.first() {
-            if let Some(group) = self.editor_groups.get(gid) {
-                ws_session.open_files = files_for_group(group);
-            }
-        }
-        if group_ids.len() >= 2 {
-            if let Some(group) = self.editor_groups.get(&group_ids[1]) {
-                ws_session.open_files_group1 = files_for_group(group);
-            }
-        }
-        ws_session.active_file = self.file_path().cloned();
-        ws_session.file_positions = self.session.file_positions.clone();
-        // Save active_group as index position in leaf order for backward compat
-        ws_session.active_group = group_ids
-            .iter()
-            .position(|&id| id == self.active_group)
-            .unwrap_or(0);
-        // For backward-compat, extract direction and ratio from root split
-        // (old format only supported a single split).
-        if let GroupLayout::Split {
-            direction, ratio, ..
-        } = &self.group_layout
-        {
-            ws_session.group_split_direction = match direction {
-                SplitDirection::Vertical => 0,
-                SplitDirection::Horizontal => 1,
-            };
-            ws_session.group_split_ratio = *ratio;
-        }
-        // Save the full recursive tree layout (new format).
-        ws_session.group_layout = Some(self.build_session_group_layout(&self.group_layout));
-        ws_session.save_for_workspace(root).ok();
-    }
-
-    /// Recursively convert the engine's GroupLayout tree into a SessionGroupLayout
-    /// for serialization, collecting each leaf group's open file paths.
-    fn build_session_group_layout(&self, layout: &GroupLayout) -> SessionGroupLayout {
-        match layout {
-            GroupLayout::Leaf(gid) => {
-                let files = self
-                    .editor_groups
-                    .get(gid)
-                    .map(|group| {
-                        let mut files: Vec<PathBuf> = Vec::new();
-                        for tab in &group.tabs {
-                            if let Some(window) = self.windows.get(&tab.active_window) {
-                                if let Some(bs) = self.buffer_manager.get(window.buffer_id) {
-                                    if let Some(ref fp) = bs.file_path {
-                                        if !files.contains(fp) {
-                                            files.push(fp.clone());
-                                        }
-                                    }
-                                }
-                            }
-                        }
-                        files
-                    })
-                    .unwrap_or_default();
-                SessionGroupLayout::Leaf { files }
-            }
-            GroupLayout::Split {
-                direction,
-                ratio,
-                first,
-                second,
-            } => SessionGroupLayout::Split {
-                direction: match direction {
-                    SplitDirection::Vertical => 0,
-                    SplitDirection::Horizontal => 1,
-                },
-                ratio: *ratio,
-                first: Box::new(self.build_session_group_layout(first)),
-                second: Box::new(self.build_session_group_layout(second)),
-            },
-        }
-    }
-
-    /// Recursively restore groups from a SessionGroupLayout tree.
-    /// Returns the reconstructed GroupLayout tree.
-    fn restore_session_group_layout(&mut self, session_layout: &SessionGroupLayout) -> GroupLayout {
-        match session_layout {
-            SessionGroupLayout::Leaf { files } => {
-                let gid = self.new_group_id();
-                // Create the group with the first file (or a scratch tab if no files).
-                let valid: Vec<&PathBuf> = files.iter().filter(|p| p.exists()).collect();
-                if valid.is_empty() {
-                    // Empty group: create a fresh scratch buffer (don't rely on active_group chain).
-                    let wid = self.new_window_id();
-                    let buf_id = self.buffer_manager.create();
-                    let w = Window::new(wid, buf_id);
-                    self.windows.insert(wid, w);
-                    let tid = self.new_tab_id();
-                    let tab = Tab::new(tid, wid);
-                    self.editor_groups.insert(gid, EditorGroup::new(tab));
-                } else {
-                    // Open files in this group's tabs.
-                    let mut first = true;
-                    for path in &valid {
-                        let wid = self.new_window_id();
-                        let buf_id = self
-                            .buffer_manager
-                            .open_file(path)
-                            .unwrap_or_else(|_| self.buffer_manager.create());
-                        let mut w = Window::new(wid, buf_id);
-                        let view = self.restore_file_position(buf_id);
-                        w.view = view;
-                        self.windows.insert(wid, w);
-                        let tid = self.new_tab_id();
-                        let tab = Tab::new(tid, wid);
-                        if first {
-                            self.editor_groups.insert(gid, EditorGroup::new(tab));
-                            first = false;
-                        } else if let Some(group) = self.editor_groups.get_mut(&gid) {
-                            group.tabs.push(tab);
-                        }
-                    }
-                }
-                GroupLayout::Leaf(gid)
-            }
-            SessionGroupLayout::Split {
-                direction,
-                ratio,
-                first,
-                second,
-            } => {
-                let dir = if *direction == 1 {
-                    SplitDirection::Horizontal
-                } else {
-                    SplitDirection::Vertical
-                };
-                let first_layout = self.restore_session_group_layout(first);
-                let second_layout = self.restore_session_group_layout(second);
-                GroupLayout::Split {
-                    direction: dir,
-                    ratio: *ratio,
-                    first: Box::new(first_layout),
-                    second: Box::new(second_layout),
-                }
-            }
-        }
-    }
-
-    // ─── User keymaps ────────────────────────────────────────────────────────
-
-    /// Rebuild the parsed user_keymaps cache from settings.keymaps.
-    /// Call after loading or changing settings.
-    pub fn rebuild_user_keymaps(&mut self) {
-        self.user_keymaps = self
-            .settings
-            .keymaps
-            .iter()
-            .filter_map(|s| parse_keymap_def(s))
-            .collect();
-    }
-
-    /// Check user keymaps for the current keypress. Returns `Some(action)` if
-    /// an exact match was found, `None` to fall through to built-in handling.
-    /// Handles multi-key sequences by buffering keypresses.
-    fn try_user_keymap(
-        &mut self,
-        key_name: &str,
-        unicode: Option<char>,
-        ctrl: bool,
-        changed: &mut bool,
-    ) -> Option<EngineAction> {
-        if self.keymap_replaying || self.user_keymaps.is_empty() {
-            return None;
-        }
-
-        let mode_str = if self.is_vscode_mode() {
-            // VSCode mode has no modal distinction; "n" keymaps apply.
-            "n"
-        } else {
-            match self.mode {
-                Mode::Normal => "n",
-                Mode::Visual | Mode::VisualLine | Mode::VisualBlock => "v",
-                Mode::Insert => "i",
-                Mode::Command => "c",
-                _ => return None,
-            }
-        };
-
-        let encoded = encode_keypress(key_name, unicode, ctrl);
-        self.keymap_buf.push(encoded);
-
-        let mut exact_match_action = None;
-        let mut has_prefix = false;
-
-        for km in &self.user_keymaps {
-            if km.mode != mode_str {
-                continue;
-            }
-            if km.keys == self.keymap_buf {
-                exact_match_action = Some(km.action.clone());
-            } else if km.keys.len() > self.keymap_buf.len()
-                && km.keys[..self.keymap_buf.len()] == self.keymap_buf[..]
-            {
-                has_prefix = true;
-            }
-        }
-
-        if let Some(action) = exact_match_action {
-            self.keymap_buf.clear();
-            let count = self.take_count();
-            // Substitute {count} in the action, or append count as argument
-            let cmd = if action.contains("{count}") {
-                action.replace("{count}", &count.to_string())
-            } else if count > 1 {
-                format!("{action} {count}")
-            } else {
-                action
-            };
-            *changed = true;
-            return Some(self.execute_command(&cmd));
-        }
-
-        if has_prefix {
-            // More keys needed — consume this keypress
-            return Some(EngineAction::None);
-        }
-
-        // No match and no prefix. Replay buffered keys.
-        let buf: Vec<String> = self.keymap_buf.drain(..).collect();
-        if buf.len() <= 1 {
-            // Single key, no match — fall through to built-in handling
-            return None;
-        }
-
-        // Multi-key sequence that didn't match any keymap: replay all keys
-        self.keymap_replaying = true;
-        let mut last_action = EngineAction::None;
-        for encoded_key in buf {
-            let (rk_name, rk_unicode, rk_ctrl) = decode_keypress(&encoded_key);
-            last_action = self.handle_key(&rk_name, rk_unicode, rk_ctrl);
-        }
-        self.keymap_replaying = false;
-        Some(last_action)
-    }
-
-    /// Try to run a named plugin command. Returns `true` if the command was found.
-    pub fn plugin_run_command(&mut self, name: &str, args: &str) -> bool {
-        if !self.settings.plugins_enabled {
-            return false;
-        }
-        let pm = match self.plugin_manager.take() {
-            Some(p) => p,
-            None => return false,
-        };
-        let ctx = self.make_plugin_ctx(false);
-        let (found, ctx) = pm.call_command(name, args, ctx);
-        self.plugin_manager = Some(pm);
-        self.apply_plugin_ctx(ctx);
-        found
-    }
-
-    /// Try to run a plugin keymap. Returns `true` if a mapping was found and executed.
-    pub fn plugin_run_keymap(&mut self, mode: &str, key: &str) -> bool {
-        if !self.settings.plugins_enabled {
-            return false;
-        }
-        let pm = match self.plugin_manager.take() {
-            Some(p) => p,
-            None => return false,
-        };
-        let ctx = self.make_plugin_ctx(false);
-        let (found, ctx) = pm.call_keymap(mode, key, ctx);
-        self.plugin_manager = Some(pm);
-        self.apply_plugin_ctx(ctx);
-        found
-    }
-
-    // =======================================================================
-    // Bracket navigation ([ and ] commands)
-    // =======================================================================
-
-    /// Jump to next section start (]] or next section end (][).
-    /// `end_section`: false = start ('{' in column 0), true = end ('}' in column 0).
-    /// In LaTeX buffers: ]] jumps to next \section/\chapter/\subsection/\subsubsection,
-    /// ][ jumps to next \end{...}.
-    fn jump_section_forward(&mut self, end_section: bool) {
-        if self.is_latex_buffer() {
-            self.jump_latex_section_forward(end_section);
-            return;
-        }
-        let target_char = if end_section { '}' } else { '{' };
-        let total = self.buffer().len_lines();
-        let start = self.view().cursor.line + 1;
-        for line in start..total {
-            let line_start = self.buffer().line_to_char(line);
-            if self.buffer().line_len_chars(line) > 0
-                && self.buffer().content.char(line_start) == target_char
-            {
-                self.view_mut().cursor.line = line;
-                self.view_mut().cursor.col = 0;
-                return;
-            }
-        }
-    }
-
-    /// Jump to previous section start ([[) or previous section end (][]).
-    /// In LaTeX buffers: [[ jumps to previous \section/etc., [] jumps to previous \end{}.
-    fn jump_section_backward(&mut self, end_section: bool) {
-        if self.is_latex_buffer() {
-            self.jump_latex_section_backward(end_section);
-            return;
-        }
-        let target_char = if end_section { '}' } else { '{' };
-        let cur = self.view().cursor.line;
-        for line in (0..cur).rev() {
-            let line_start = self.buffer().line_to_char(line);
-            if self.buffer().line_len_chars(line) > 0
-                && self.buffer().content.char(line_start) == target_char
-            {
-                self.view_mut().cursor.line = line;
-                self.view_mut().cursor.col = 0;
-                return;
-            }
-        }
-    }
-
-    /// Jump to next method start (]m) — finds next '{' that starts a block.
-    /// In LaTeX buffers: ]m jumps to next \begin{...}.
-    fn jump_method_start_forward(&mut self) {
-        if self.is_latex_buffer() {
-            self.jump_latex_env_forward(false);
-            return;
-        }
-        let total_chars = self.buffer().len_chars();
-        let cur_pos = self.buffer().line_to_char(self.view().cursor.line) + self.view().cursor.col;
-        let mut pos = cur_pos + 1;
-        while pos < total_chars {
-            if self.buffer().content.char(pos) == '{' {
-                let line = self.buffer().content.char_to_line(pos);
-                let line_start = self.buffer().line_to_char(line);
-                self.view_mut().cursor.line = line;
-                self.view_mut().cursor.col = pos - line_start;
-                return;
-            }
-            pos += 1;
-        }
-    }
-
-    /// Jump to previous method start ([m).
-    /// In LaTeX buffers: [m jumps to previous \begin{...}.
-    fn jump_method_start_backward(&mut self) {
-        if self.is_latex_buffer() {
-            self.jump_latex_env_backward(false);
-            return;
-        }
-        let cur_pos = self.buffer().line_to_char(self.view().cursor.line) + self.view().cursor.col;
-        if cur_pos == 0 {
-            return;
-        }
-        let mut pos = cur_pos - 1;
-        loop {
-            if self.buffer().content.char(pos) == '{' {
-                let line = self.buffer().content.char_to_line(pos);
-                let line_start = self.buffer().line_to_char(line);
-                self.view_mut().cursor.line = line;
-                self.view_mut().cursor.col = pos - line_start;
-                return;
-            }
-            if pos == 0 {
-                break;
-            }
-            pos -= 1;
-        }
-    }
-
-    /// Jump to next method end (]M) — finds next '}'.
-    /// In LaTeX buffers: ]M jumps to next \end{...}.
-    fn jump_method_end_forward(&mut self) {
-        if self.is_latex_buffer() {
-            self.jump_latex_env_forward(true);
-            return;
-        }
-        let total_chars = self.buffer().len_chars();
-        let cur_pos = self.buffer().line_to_char(self.view().cursor.line) + self.view().cursor.col;
-        let mut pos = cur_pos + 1;
-        while pos < total_chars {
-            if self.buffer().content.char(pos) == '}' {
-                let line = self.buffer().content.char_to_line(pos);
-                let line_start = self.buffer().line_to_char(line);
-                self.view_mut().cursor.line = line;
-                self.view_mut().cursor.col = pos - line_start;
-                return;
-            }
-            pos += 1;
-        }
-    }
-
-    /// Jump to previous method end ([M).
-    /// In LaTeX buffers: [M jumps to previous \end{...}.
-    fn jump_method_end_backward(&mut self) {
-        if self.is_latex_buffer() {
-            self.jump_latex_env_backward(true);
-            return;
-        }
-        let cur_pos = self.buffer().line_to_char(self.view().cursor.line) + self.view().cursor.col;
-        if cur_pos == 0 {
-            return;
-        }
-        let mut pos = cur_pos - 1;
-        loop {
-            if self.buffer().content.char(pos) == '}' {
-                let line = self.buffer().content.char_to_line(pos);
-                let line_start = self.buffer().line_to_char(line);
-                self.view_mut().cursor.line = line;
-                self.view_mut().cursor.col = pos - line_start;
-                return;
-            }
-            if pos == 0 {
-                break;
-            }
-            pos -= 1;
-        }
-    }
-
-    // --- LaTeX-specific motion helpers ---
-
-    /// LaTeX section commands to match for ]] / [[ jumps.
-    const LATEX_SECTION_COMMANDS: &'static [&'static str] = &[
-        "\\part",
-        "\\chapter",
-        "\\section",
-        "\\subsection",
-        "\\subsubsection",
-        "\\paragraph",
-        "\\subparagraph",
-    ];
-
-    /// Jump forward to next LaTeX section command (]]) or \end{} (][).
-    fn jump_latex_section_forward(&mut self, end_section: bool) {
-        let total = self.buffer().len_lines();
-        let start = self.view().cursor.line + 1;
-        for line in start..total {
-            let line_text = self.buffer().content.line(line).chars().collect::<String>();
-            let trimmed = line_text.trim_start();
-            if end_section {
-                if trimmed.starts_with("\\end{") {
-                    self.view_mut().cursor.line = line;
-                    self.view_mut().cursor.col = 0;
-                    return;
-                }
-            } else {
-                for cmd in Self::LATEX_SECTION_COMMANDS {
-                    if let Some(after) = trimmed.strip_prefix(cmd) {
-                        if after.starts_with('{') || after.starts_with('*') || after.is_empty() {
-                            self.view_mut().cursor.line = line;
-                            self.view_mut().cursor.col = 0;
-                            return;
-                        }
-                    }
-                }
-            }
-        }
-    }
-
-    /// Jump backward to previous LaTeX section command ([[) or \end{} ([]).
-    fn jump_latex_section_backward(&mut self, end_section: bool) {
-        let cur = self.view().cursor.line;
-        for line in (0..cur).rev() {
-            let line_text = self.buffer().content.line(line).chars().collect::<String>();
-            let trimmed = line_text.trim_start();
-            if end_section {
-                if trimmed.starts_with("\\end{") {
-                    self.view_mut().cursor.line = line;
-                    self.view_mut().cursor.col = 0;
-                    return;
-                }
-            } else {
-                for cmd in Self::LATEX_SECTION_COMMANDS {
-                    if let Some(after) = trimmed.strip_prefix(cmd) {
-                        if after.starts_with('{') || after.starts_with('*') || after.is_empty() {
-                            self.view_mut().cursor.line = line;
-                            self.view_mut().cursor.col = 0;
-                            return;
-                        }
-                    }
-                }
-            }
-        }
-    }
-
-    /// Jump forward to next \begin{} (is_end=false) or \end{} (is_end=true).
-    fn jump_latex_env_forward(&mut self, is_end: bool) {
-        let needle = if is_end { "\\end{" } else { "\\begin{" };
-        let total = self.buffer().len_lines();
-        let start_line = self.view().cursor.line;
-        let start_col = self.view().cursor.col;
-        for line in start_line..total {
-            let line_text = self.buffer().content.line(line).chars().collect::<String>();
-            let search_from = if line == start_line { start_col + 1 } else { 0 };
-            if search_from < line_text.len() {
-                if let Some(rel) = line_text[search_from..].find(needle) {
-                    self.view_mut().cursor.line = line;
-                    self.view_mut().cursor.col = search_from + rel;
-                    return;
-                }
-            }
-        }
-    }
-
-    /// Jump backward to previous \begin{} (is_end=false) or \end{} (is_end=true).
-    fn jump_latex_env_backward(&mut self, is_end: bool) {
-        let needle = if is_end { "\\end{" } else { "\\begin{" };
-        let start_line = self.view().cursor.line;
-        let start_col = self.view().cursor.col;
-        for line in (0..=start_line).rev() {
-            let line_text = self.buffer().content.line(line).chars().collect::<String>();
-            let search_end = if line == start_line {
-                start_col
-            } else {
-                line_text.len()
-            };
-            if let Some(pos) = line_text[..search_end].rfind(needle) {
-                self.view_mut().cursor.line = line;
-                self.view_mut().cursor.col = pos;
-                return;
-            }
-        }
-    }
-
-    /// Jump forward to next unmatched close bracket (]} or ])).
-    fn jump_unmatched_forward(&mut self, open: char, close: char) {
-        let total_chars = self.buffer().len_chars();
-        let cur_pos = self.buffer().line_to_char(self.view().cursor.line) + self.view().cursor.col;
-        let mut pos = cur_pos + 1;
-        let mut depth: i32 = 0;
-        while pos < total_chars {
-            let ch = self.buffer().content.char(pos);
-            if ch == open {
-                depth += 1;
-            } else if ch == close {
-                if depth == 0 {
-                    let line = self.buffer().content.char_to_line(pos);
-                    let line_start = self.buffer().line_to_char(line);
-                    self.view_mut().cursor.line = line;
-                    self.view_mut().cursor.col = pos - line_start;
-                    return;
-                }
-                depth -= 1;
-            }
-            pos += 1;
-        }
-    }
-
-    /// Jump backward to previous unmatched open bracket ([{ or [().
-    fn jump_unmatched_backward(&mut self, open: char, close: char) {
-        let cur_pos = self.buffer().line_to_char(self.view().cursor.line) + self.view().cursor.col;
-        if cur_pos == 0 {
-            return;
-        }
-        let mut pos = cur_pos - 1;
-        let mut depth: i32 = 0;
-        loop {
-            let ch = self.buffer().content.char(pos);
-            if ch == close {
-                depth += 1;
-            } else if ch == open {
-                if depth == 0 {
-                    let line = self.buffer().content.char_to_line(pos);
-                    let line_start = self.buffer().line_to_char(line);
-                    self.view_mut().cursor.line = line;
-                    self.view_mut().cursor.col = pos - line_start;
-                    return;
-                }
-                depth -= 1;
-            }
-            if pos == 0 {
-                break;
-            }
-            pos -= 1;
-        }
-    }
-
-    // =======================================================================
-    // Window resize (CTRL-W +/-/</>=/|/_)
-    // =======================================================================
-
-    /// Resize the window's parent split by delta steps.
-    /// `direction`: which split direction to look for (Horizontal for +/-, Vertical for </>).
-    /// `increase`: true = make active group bigger, false = smaller.
-    fn resize_window_split(&mut self, direction: SplitDirection, increase: bool, count: usize) {
-        let delta_per_step = 0.05;
-        if let Some((split_idx, split_dir, is_first)) =
-            self.group_layout.parent_split_of(self.active_group)
-        {
-            if split_dir == direction {
-                // Active group is in first child → increasing ratio makes it bigger
-                let delta = if (is_first && increase) || (!is_first && !increase) {
-                    delta_per_step * count as f64
-                } else {
-                    -(delta_per_step * count as f64)
-                };
-                self.group_layout.adjust_ratio_at_index(split_idx, delta);
-            }
-        }
-    }
-
-    /// Equalize all split ratios to 0.5.
-    fn equalize_splits(&mut self) {
-        self.group_layout.set_all_ratios(0.5);
-    }
-
-    /// Maximize window in a given direction (CTRL-W _ for height, CTRL-W | for width).
-    fn maximize_window_split(&mut self, direction: SplitDirection) {
-        if let Some((split_idx, split_dir, is_first)) =
-            self.group_layout.parent_split_of(self.active_group)
-        {
-            if split_dir == direction {
-                let ratio = if is_first { 0.9 } else { 0.1 };
-                self.group_layout.set_ratio_at_index(split_idx, ratio);
-            }
-        }
-    }
-
-    /// Execute a window command by character (`:wincmd {char}` and Ctrl-W {char}).
-    fn execute_wincmd(&mut self, ch: char, count: usize) -> EngineAction {
-        match ch {
-            // Focus
-            'h' => self.focus_window_direction(SplitDirection::Vertical, false),
-            'j' => self.focus_window_direction(SplitDirection::Horizontal, true),
-            'k' => self.focus_window_direction(SplitDirection::Horizontal, false),
-            'l' => self.focus_window_direction(SplitDirection::Vertical, true),
-            'w' | 'W' => self.focus_next_window(),
-            'p' => {
-                if let Some(prev) = self.prev_active_group {
-                    if self.editor_groups.contains_key(&prev) {
-                        let cur = self.active_group;
-                        self.active_group = prev;
-                        self.prev_active_group = Some(cur);
-                    }
-                }
-            }
-            't' => {
-                if let Some(first) = self.group_layout.nth_leaf(0) {
-                    if first != self.active_group {
-                        self.prev_active_group = Some(self.active_group);
-                    }
-                    self.active_group = first;
-                }
-            }
-            'b' => {
-                let ids = self.group_layout.group_ids();
-                if let Some(&last) = ids.last() {
-                    if last != self.active_group {
-                        self.prev_active_group = Some(self.active_group);
-                    }
-                    self.active_group = last;
-                }
-            }
-            // Move
-            'H' => self.move_window_to_edge(SplitDirection::Vertical, false),
-            'J' => self.move_window_to_edge(SplitDirection::Horizontal, true),
-            'K' => self.move_window_to_edge(SplitDirection::Horizontal, false),
-            'L' => self.move_window_to_edge(SplitDirection::Vertical, true),
-            'T' => self.move_window_to_new_group(),
-            'x' => self.exchange_windows(),
-            'r' => self.rotate_windows(true),
-            'R' => self.rotate_windows(false),
-            // Split / Close
-            's' | 'S' => self.split_window(SplitDirection::Horizontal, None),
-            'v' | 'V' => self.split_window(SplitDirection::Vertical, None),
-            'c' | 'C' => {
-                self.close_window();
-            }
-            'q' => {
-                self.close_window();
-            }
-            'o' | 'O' => self.close_other_windows(),
-            'n' => {
-                let _ = self.execute_command("new");
-            }
-            // Editor groups
-            'e' => self.open_editor_group(SplitDirection::Vertical),
-            'E' => self.open_editor_group(SplitDirection::Horizontal),
-            // Resize (count-aware)
-            '+' => self.resize_window_split(SplitDirection::Horizontal, true, count),
-            '-' => self.resize_window_split(SplitDirection::Horizontal, false, count),
-            '>' => self.resize_window_split(SplitDirection::Vertical, true, count),
-            '<' => self.resize_window_split(SplitDirection::Vertical, false, count),
-            '=' => self.equalize_splits(),
-            '_' => self.maximize_window_split(SplitDirection::Horizontal),
-            '|' => self.maximize_window_split(SplitDirection::Vertical),
-            // Composite
-            'f' => {
-                if let Some(path) = self.file_path_under_cursor() {
-                    let abs_path = if path.is_absolute() {
-                        path
-                    } else {
-                        self.cwd.join(&path)
-                    };
-                    self.split_window(SplitDirection::Horizontal, None);
-                    return EngineAction::OpenFile(abs_path);
-                } else {
-                    self.message = "No file path under cursor".to_string();
-                }
-            }
-            'd' => {
-                self.split_window(SplitDirection::Horizontal, None);
-                self.push_jump_location();
-                self.lsp_request_definition();
-            }
-            _ => {
-                self.message = format!("Unknown wincmd: {}", ch);
-            }
-        }
-        EngineAction::None
-    }
-
-    /// Ctrl-W H/J/K/L: move current window to far edge.
-    /// Creates a new editor group at the edge of the entire layout.
-    fn move_window_to_edge(&mut self, direction: SplitDirection, forward: bool) {
-        // Only meaningful with multiple groups
-        let groups = self.group_layout.group_ids();
-        if groups.len() <= 1 {
-            // With a single group, split the group layout at the root
-            let buf_id = self.active_buffer_id();
-            let view_clone = self.view().clone();
-            // Close current window if possible
-            let could_close = self.close_window();
-            if !could_close {
-                return; // Last window, can't move
-            }
-            // Create new group at the edge
-            let new_win_id = self.new_window_id();
-            let mut new_win = Window::new(new_win_id, buf_id);
-            new_win.view = view_clone;
-            self.windows.insert(new_win_id, new_win);
-            let tab = Tab::new(self.new_tab_id(), new_win_id);
-            let new_gid = self.new_group_id();
-            self.editor_groups.insert(new_gid, EditorGroup::new(tab));
-            // Wrap the existing layout in a split with the new group at the desired edge
-            let old_layout =
-                std::mem::replace(&mut self.group_layout, GroupLayout::leaf(GroupId(0)));
-            self.group_layout = if forward {
-                GroupLayout::Split {
-                    direction,
-                    ratio: 0.5,
-                    first: Box::new(old_layout),
-                    second: Box::new(GroupLayout::leaf(new_gid)),
-                }
-            } else {
-                GroupLayout::Split {
-                    direction,
-                    ratio: 0.5,
-                    first: Box::new(GroupLayout::leaf(new_gid)),
-                    second: Box::new(old_layout),
-                }
-            };
-            self.prev_active_group = Some(self.active_group);
-            self.active_group = new_gid;
-            return;
-        }
-        // Multiple groups: remove window, create new group at edge
-        let buf_id = self.active_buffer_id();
-        let view_clone = self.view().clone();
-        let could_close = self.close_window();
-        if !could_close {
-            return;
-        }
-        let new_win_id = self.new_window_id();
-        let mut new_win = Window::new(new_win_id, buf_id);
-        new_win.view = view_clone;
-        self.windows.insert(new_win_id, new_win);
-        let tab = Tab::new(self.new_tab_id(), new_win_id);
-        let new_gid = self.new_group_id();
-        self.editor_groups.insert(new_gid, EditorGroup::new(tab));
-        let old_layout = std::mem::replace(&mut self.group_layout, GroupLayout::leaf(GroupId(0)));
-        self.group_layout = if forward {
-            GroupLayout::Split {
-                direction,
-                ratio: 0.5,
-                first: Box::new(old_layout),
-                second: Box::new(GroupLayout::leaf(new_gid)),
-            }
-        } else {
-            GroupLayout::Split {
-                direction,
-                ratio: 0.5,
-                first: Box::new(GroupLayout::leaf(new_gid)),
-                second: Box::new(old_layout),
-            }
-        };
-        self.prev_active_group = Some(self.active_group);
-        self.active_group = new_gid;
-    }
-
-    /// Ctrl-W T: move current window to a new editor group.
-    fn move_window_to_new_group(&mut self) {
-        // Only meaningful with multiple windows in the current tab
-        let tab = self.active_tab();
-        if tab.layout.is_single_window() && self.active_group().tabs.len() == 1 {
-            self.message = "Already the only window".to_string();
-            return;
-        }
-        let buf_id = self.active_buffer_id();
-        let view_clone = self.view().clone();
-        let could_close = self.close_window();
-        if !could_close {
-            return;
-        }
-        // Open a new editor group with that buffer
-        let new_win_id = self.new_window_id();
-        let mut new_win = Window::new(new_win_id, buf_id);
-        new_win.view = view_clone;
-        self.windows.insert(new_win_id, new_win);
-        let tab = Tab::new(self.new_tab_id(), new_win_id);
-        let new_gid = self.new_group_id();
-        self.editor_groups.insert(new_gid, EditorGroup::new(tab));
-        self.group_layout
-            .split_at(self.active_group, SplitDirection::Vertical, new_gid, false);
-        self.prev_active_group = Some(self.active_group);
-        self.active_group = new_gid;
-    }
-
-    /// Ctrl-W x: exchange current window with next window in the same tab.
-    fn exchange_windows(&mut self) {
-        let tab = self.active_tab();
-        let ids = tab.layout.window_ids();
-        if ids.len() < 2 {
-            return;
-        }
-        let current_id = tab.active_window;
-        let current_idx = ids.iter().position(|&id| id == current_id).unwrap_or(0);
-        let next_idx = (current_idx + 1) % ids.len();
-        let next_id = ids[next_idx];
-        // Swap buffer_id and view between the two windows
-        let current_buf = self.windows[&current_id].buffer_id;
-        let current_view = self.windows[&current_id].view.clone();
-        let next_buf = self.windows[&next_id].buffer_id;
-        let next_view = self.windows[&next_id].view.clone();
-        if let Some(w) = self.windows.get_mut(&current_id) {
-            w.buffer_id = next_buf;
-            w.view = next_view;
-        }
-        if let Some(w) = self.windows.get_mut(&next_id) {
-            w.buffer_id = current_buf;
-            w.view = current_view;
-        }
-    }
-
-    /// Ctrl-W r/R: rotate windows in the current tab.
-    /// `forward=true` rotates downward/rightward, `forward=false` rotates upward/leftward.
-    fn rotate_windows(&mut self, forward: bool) {
-        let tab = self.active_tab();
-        let ids = tab.layout.window_ids();
-        if ids.len() < 2 {
-            return;
-        }
-        // Collect (buffer_id, view) for each window in layout order
-        let mut data: Vec<_> = ids
-            .iter()
-            .map(|&id| {
-                let w = &self.windows[&id];
-                (w.buffer_id, w.view.clone())
-            })
-            .collect();
-        // Rotate the data
-        if forward {
-            // Last element moves to front
-            let last = data.pop().unwrap();
-            data.insert(0, last);
-        } else {
-            // First element moves to back
-            let first = data.remove(0);
-            data.push(first);
-        }
-        // Apply rotated data back
-        for (i, &id) in ids.iter().enumerate() {
-            if let Some(w) = self.windows.get_mut(&id) {
-                w.buffer_id = data[i].0;
-                w.view = data[i].1.clone();
-            }
-        }
-    }
-
-    /// Jump to end of C-style comment block (]*  or  ]/).
-    fn jump_comment_end(&mut self) {
-        let total = self.buffer().len_lines();
-        let start = self.view().cursor.line + 1;
-        for line_idx in start..total {
-            let line: String = self.buffer().content.line(line_idx).chars().collect();
-            let trimmed = line.trim();
-            if trimmed.contains("*/") {
-                self.view_mut().cursor.line = line_idx;
-                // Position cursor at the '*' of '*/'
-                if let Some(pos) = line.find("*/") {
-                    let col = line[..pos].chars().count();
-                    self.view_mut().cursor.col = col;
-                } else {
-                    self.view_mut().cursor.col = 0;
-                }
-                self.clamp_cursor_col();
-                return;
-            }
-        }
-    }
-
-    /// Jump to start of C-style comment block ([*  or  [/).
-    fn jump_comment_start(&mut self) {
-        let cursor_line = self.view().cursor.line;
-        if cursor_line == 0 {
-            return;
-        }
-        for line_idx in (0..cursor_line).rev() {
-            let line: String = self.buffer().content.line(line_idx).chars().collect();
-            let trimmed = line.trim();
-            if trimmed.contains("/*") {
-                self.view_mut().cursor.line = line_idx;
-                // Position cursor at the '/' of '/*'
-                if let Some(pos) = line.find("/*") {
-                    let col = line[..pos].chars().count();
-                    self.view_mut().cursor.col = col;
-                } else {
-                    self.view_mut().cursor.col = 0;
-                }
-                self.clamp_cursor_col();
-                return;
-            }
-        }
-    }
-
-    /// Jump forward to next unmatched `#else` or `#endif` (`]#`).
-    /// Uses depth tracking: `#if`/`#ifdef`/`#ifndef` increase depth,
-    /// `#endif` decreases depth, `#else`/`#elif` match at depth 0.
-    fn jump_preproc_forward(&mut self) {
-        let total = self.buffer().len_lines();
-        let start = self.view().cursor.line + 1;
-        let mut depth: i32 = 0;
-        for line_idx in start..total {
-            let line: String = self.buffer().content.line(line_idx).chars().collect();
-            let trimmed = line.trim();
-            if let Some(directive) = Self::preproc_directive(trimmed) {
-                match directive {
-                    PreprocKind::If => depth += 1,
-                    PreprocKind::ElseElif => {
-                        if depth == 0 {
-                            self.view_mut().cursor.line = line_idx;
-                            self.view_mut().cursor.col = 0;
-                            self.clamp_cursor_col();
-                            return;
-                        }
-                    }
-                    PreprocKind::Endif => {
-                        if depth == 0 {
-                            self.view_mut().cursor.line = line_idx;
-                            self.view_mut().cursor.col = 0;
-                            self.clamp_cursor_col();
-                            return;
-                        }
-                        depth -= 1;
-                    }
-                }
-            }
-        }
-    }
-
-    /// Jump backward to previous unmatched `#if` or `#else` (`[#`).
-    /// Uses depth tracking: `#endif` increases depth,
-    /// `#if`/`#ifdef`/`#ifndef` decrease depth, `#else`/`#elif` match at depth 0.
-    fn jump_preproc_backward(&mut self) {
-        let cursor_line = self.view().cursor.line;
-        if cursor_line == 0 {
-            return;
-        }
-        let mut depth: i32 = 0;
-        for line_idx in (0..cursor_line).rev() {
-            let line: String = self.buffer().content.line(line_idx).chars().collect();
-            let trimmed = line.trim();
-            if let Some(directive) = Self::preproc_directive(trimmed) {
-                match directive {
-                    PreprocKind::Endif => depth += 1,
-                    PreprocKind::ElseElif => {
-                        if depth == 0 {
-                            self.view_mut().cursor.line = line_idx;
-                            self.view_mut().cursor.col = 0;
-                            self.clamp_cursor_col();
-                            return;
-                        }
-                    }
-                    PreprocKind::If => {
-                        if depth == 0 {
-                            self.view_mut().cursor.line = line_idx;
-                            self.view_mut().cursor.col = 0;
-                            self.clamp_cursor_col();
-                            return;
-                        }
-                        depth -= 1;
-                    }
-                }
-            }
-        }
-    }
-
-    /// Classify a trimmed line as a preprocessor directive kind.
-    fn preproc_directive(trimmed: &str) -> Option<PreprocKind> {
-        if !trimmed.starts_with('#') {
-            return None;
-        }
-        // Strip '#' and optional whitespace after it
-        let after_hash = trimmed[1..].trim_start();
-        if after_hash.starts_with("ifdef")
-            || after_hash.starts_with("ifndef")
-            || after_hash.starts_with("if ")
-            || after_hash.starts_with("if\t")
-            || after_hash == "if"
-        {
-            Some(PreprocKind::If)
-        } else if after_hash.starts_with("else") || after_hash.starts_with("elif") {
-            Some(PreprocKind::ElseElif)
-        } else if after_hash.starts_with("endif") {
-            Some(PreprocKind::Endif)
-        } else {
-            None
-        }
-    }
-
-    /// `do` (diff obtain): in a diff view, replace the current line in the active
-    /// window with the corresponding line from the other diff window.
-    fn diff_obtain(&mut self, changed: &mut bool) {
-        let (a_win, b_win) = match self.diff_window_pair {
-            Some(pair) => pair,
-            None => {
-                self.message = "Not in diff mode".to_string();
-                return;
-            }
-        };
-        let active = self.active_window_id();
-        let other = if active == a_win {
-            b_win
-        } else if active == b_win {
-            a_win
-        } else {
-            self.message = "Current window is not part of a diff".to_string();
-            return;
-        };
-        let cursor_line = self.view().cursor.line;
-        // Get the diff status for the active window
-        let diff_status = self
-            .diff_results
-            .get(&active)
-            .and_then(|v| v.get(cursor_line))
-            .cloned();
-        match diff_status {
-            Some(DiffLine::Same) => {
-                self.message = "Line is the same in both files".to_string();
-            }
-            Some(DiffLine::Added) | Some(DiffLine::Removed) | Some(DiffLine::Padding) | None => {
-                // Get the corresponding line from the other window
-                // For simplicity, use the same line number from the other buffer
-                if let Some(other_win) = self.windows.get(&other) {
-                    let other_buf_id = other_win.buffer_id;
-                    if let Some(other_state) = self.buffer_manager.get(other_buf_id) {
-                        if cursor_line < other_state.buffer.len_lines() {
-                            let other_line: String = other_state
-                                .buffer
-                                .content
-                                .line(cursor_line)
-                                .chars()
-                                .collect();
-                            // Replace the current line
-                            let line_start = self.buffer().line_to_char(cursor_line);
-                            let line_end = if cursor_line + 1 < self.buffer().len_lines() {
-                                self.buffer().line_to_char(cursor_line + 1)
-                            } else {
-                                self.buffer().len_chars()
-                            };
-                            self.start_undo_group();
-                            self.delete_with_undo(line_start, line_end);
-                            self.insert_with_undo(line_start, &other_line);
-                            self.finish_undo_group();
-                            self.view_mut().cursor.col = 0;
-                            self.clamp_cursor_col();
-                            *changed = true;
-                            self.compute_diff();
-                        } else {
-                            self.message = "No corresponding line in other file".to_string();
-                        }
-                    }
-                }
-            }
-        }
-    }
-
-    /// `dp` (diff put): in a diff view, replace the corresponding line in the
-    /// other diff window with the current line from the active window.
-    fn diff_put(&mut self, changed: &mut bool) {
-        let (a_win, b_win) = match self.diff_window_pair {
-            Some(pair) => pair,
-            None => {
-                self.message = "Not in diff mode".to_string();
-                return;
-            }
-        };
-        let active = self.active_window_id();
-        let other = if active == a_win {
-            b_win
-        } else if active == b_win {
-            a_win
-        } else {
-            self.message = "Current window is not part of a diff".to_string();
-            return;
-        };
-        let cursor_line = self.view().cursor.line;
-        // Get the current line from the active buffer
-        let current_line: String = self.buffer().content.line(cursor_line).chars().collect();
-        // Replace the corresponding line in the other buffer
-        if let Some(other_win) = self.windows.get(&other) {
-            let other_buf_id = other_win.buffer_id;
-            if let Some(other_state) = self.buffer_manager.get_mut(other_buf_id) {
-                if cursor_line < other_state.buffer.len_lines() {
-                    let line_start = other_state.buffer.line_to_char(cursor_line);
-                    let line_end = if cursor_line + 1 < other_state.buffer.len_lines() {
-                        other_state.buffer.line_to_char(cursor_line + 1)
-                    } else {
-                        other_state.buffer.len_chars()
-                    };
-                    other_state.buffer.delete_range(line_start, line_end);
-                    other_state.buffer.insert(line_start, &current_line);
-                    other_state.dirty = true;
-                    *changed = true;
-                    self.compute_diff();
-                } else {
-                    // Other buffer is shorter — append the line
-                    let end = other_state.buffer.len_chars();
-                    let needs_newline = end > 0 && other_state.buffer.content.char(end - 1) != '\n';
-                    if needs_newline {
-                        other_state.buffer.insert(end, "\n");
-                    }
-                    let end = other_state.buffer.len_chars();
-                    other_state.buffer.insert(end, &current_line);
-                    other_state.dirty = true;
-                    *changed = true;
-                    self.compute_diff();
-                }
-            }
-        }
-    }
-
-    /// Apply an operator in blockwise mode (rectangle region).
-    #[allow(clippy::too_many_arguments)]
-    fn apply_blockwise_operator(
-        &mut self,
-        operator: char,
-        start_line: usize,
-        end_line: usize,
-        left_col: usize,
-        right_col: usize,
-        changed: &mut bool,
-    ) {
-        match operator {
-            'd' => {
-                // Delete the block region
-                self.start_undo_group();
-                let mut deleted_text = String::new();
-                // Process lines in reverse to keep indices stable
-                for line_idx in (start_line..=end_line).rev() {
-                    let line_start = self.buffer().line_to_char(line_idx);
-                    let line_len = self.buffer().line_len_chars(line_idx);
-                    let text_len = if line_len > 0
-                        && self.buffer().content.char(line_start + line_len - 1) == '\n'
-                    {
-                        line_len - 1
-                    } else {
-                        line_len
-                    };
-                    let from = left_col.min(text_len);
-                    let to = (right_col + 1).min(text_len);
-                    if from < to {
-                        let del: String = self
-                            .buffer()
-                            .content
-                            .slice((line_start + from)..(line_start + to))
-                            .chars()
-                            .collect();
-                        deleted_text = del + "\n" + &deleted_text;
-                        self.delete_with_undo(line_start + from, line_start + to);
-                    }
-                }
-                let reg = self.active_register();
-                self.set_delete_register(reg, deleted_text, false);
-                self.clear_selected_register();
-                self.view_mut().cursor.line = start_line;
-                self.view_mut().cursor.col = left_col;
-                self.clamp_cursor_col();
-                self.finish_undo_group();
-                *changed = true;
-            }
-            'y' => {
-                // Yank the block region
-                let mut yanked = String::new();
-                for line_idx in start_line..=end_line {
-                    let line_start = self.buffer().line_to_char(line_idx);
-                    let line_len = self.buffer().line_len_chars(line_idx);
-                    let text_len = if line_len > 0
-                        && self.buffer().content.char(line_start + line_len - 1) == '\n'
-                    {
-                        line_len - 1
-                    } else {
-                        line_len
-                    };
-                    let from = left_col.min(text_len);
-                    let to = (right_col + 1).min(text_len);
-                    if from < to {
-                        let chunk: String = self
-                            .buffer()
-                            .content
-                            .slice((line_start + from)..(line_start + to))
-                            .chars()
-                            .collect();
-                        yanked.push_str(&chunk);
-                    }
-                    yanked.push('\n');
-                }
-                let reg = self.active_register();
-                self.set_yank_register(reg, yanked, false);
-                self.clear_selected_register();
-            }
-            _ => {
-                // For other operators (c, ~, u, U, etc.), fall back to charwise
-                let start = self.buffer().line_to_char(start_line) + left_col;
-                let end_char = self.buffer().line_to_char(end_line) + right_col + 1;
-                let max = self.buffer().len_chars();
-                self.apply_charwise_operator(operator, start, end_char.min(max), changed);
-            }
-        }
-    }
-
-    /// Try to parse and execute a range filter command like `1,5!sort` or `.!cmd`.
-    /// Returns Some(action) if it matched, None otherwise.
-    fn try_execute_filter_command(&mut self, cmd: &str) -> Option<EngineAction> {
-        // Match patterns: N,M!cmd  or  .!cmd  or  .,.+N!cmd
-        // Split on '!' — if there's a range before and a command after, it's a filter.
-        let bang_pos = cmd.find('!')?;
-        let range_str = &cmd[..bang_pos];
-        let filter_cmd = cmd[bang_pos + 1..].trim();
-        if filter_cmd.is_empty() || range_str.is_empty() {
-            return None;
-        }
-        // Parse the range. Support: N,M  .,.+N  N  .  %
-        let (start_line, end_line) = self.parse_simple_range(range_str)?;
-        // Extract the text from the range
-        let total_lines = self.buffer().len_lines();
-        let start = start_line.min(total_lines.saturating_sub(1));
-        let end = end_line.min(total_lines.saturating_sub(1));
-        let mut lines_text = String::new();
-        for i in start..=end {
-            let line: String = self.buffer().content.line(i).chars().collect();
-            lines_text.push_str(&line);
-        }
-        // Pipe through the command
-        #[cfg(not(test))]
-        let result = {
-            use std::io::Write;
-            let mut child = match std::process::Command::new("sh")
-                .arg("-c")
-                .arg(filter_cmd)
-                .stdin(std::process::Stdio::piped())
-                .stdout(std::process::Stdio::piped())
-                .stderr(std::process::Stdio::piped())
-                .spawn()
-            {
-                Ok(c) => c,
-                Err(e) => {
-                    self.message = format!("Filter error: {}", e);
-                    return Some(EngineAction::None);
-                }
-            };
-            if let Some(mut stdin) = child.stdin.take() {
-                let _ = stdin.write_all(lines_text.as_bytes());
-            }
-            match child.wait_with_output() {
-                Ok(output) => {
-                    let stdout = String::from_utf8_lossy(&output.stdout).to_string();
-                    let stderr = String::from_utf8_lossy(&output.stderr);
-                    if !stderr.is_empty() && stdout.is_empty() {
-                        self.message = format!("Filter error: {}", stderr.trim());
-                        return Some(EngineAction::None);
-                    }
-                    stdout
-                }
-                Err(e) => {
-                    self.message = format!("Filter error: {}", e);
-                    return Some(EngineAction::None);
-                }
-            }
-        };
-        #[cfg(test)]
-        let result = lines_text.clone(); // No-op in tests
-
-        // Replace the range with the result
-        let range_start = self.buffer().line_to_char(start);
-        let range_end = if end + 1 < total_lines {
-            self.buffer().line_to_char(end + 1)
-        } else {
-            self.buffer().len_chars()
-        };
-        self.start_undo_group();
-        self.delete_with_undo(range_start, range_end);
-        self.insert_with_undo(range_start, &result);
-        self.finish_undo_group();
-        self.view_mut().cursor.line = start;
-        self.view_mut().cursor.col = 0;
-        let line_count = result.lines().count();
-        self.message = format!("{} lines filtered", line_count);
-        Some(EngineAction::None)
-    }
-
-    /// Parse a simple line range like "1,5", ".", ".,.+3", "%".
-    /// Returns 0-indexed (start_line, end_line).
-    fn parse_simple_range(&self, range: &str) -> Option<(usize, usize)> {
-        let current_line = self.view().cursor.line;
-        let last_line = self.buffer().len_lines().saturating_sub(1);
-
-        if range == "%" {
-            return Some((0, last_line));
-        }
-        if range == "." {
-            return Some((current_line, current_line));
-        }
-
-        if let Some((left, right)) = range.split_once(',') {
-            let start = self.parse_line_addr(left.trim(), current_line, last_line)?;
-            let end = self.parse_line_addr(right.trim(), current_line, last_line)?;
-            Some((start, end))
-        } else {
-            let line = self.parse_line_addr(range.trim(), current_line, last_line)?;
-            Some((line, line))
-        }
-    }
-
-    /// Parse a single line address: number (1-indexed), ".", "$", ".+N", ".-N".
-    fn parse_line_addr(&self, addr: &str, current: usize, last: usize) -> Option<usize> {
-        if addr == "." {
-            return Some(current);
-        }
-        if addr == "$" {
-            return Some(last);
-        }
-        if let Some(offset) = addr.strip_prefix(".+") {
-            let n: usize = offset.parse().ok()?;
-            return Some((current + n).min(last));
-        }
-        if let Some(offset) = addr.strip_prefix(".-") {
-            let n: usize = offset.parse().ok()?;
-            return Some(current.saturating_sub(n));
-        }
-        // Plain number (1-indexed)
-        let n: usize = addr.parse().ok()?;
-        if n == 0 {
-            return None;
-        }
-        Some((n - 1).min(last))
-    }
-
-    // =======================================================================
-    // Mouse selection (called by UI backends after coordinate conversion)
-    // =======================================================================
-
-    /// Handle a single mouse click at the given buffer position.
-    /// Exits visual mode if active, positions cursor, clears drag state.
-    pub fn mouse_click(&mut self, window_id: WindowId, line: usize, col: usize) {
-        // Exit visual mode if active
-        if matches!(
-            self.mode,
-            Mode::Visual | Mode::VisualLine | Mode::VisualBlock
-        ) {
-            self.mode = Mode::Normal;
-            self.visual_anchor = None;
-        }
-        self.mouse_drag_word_mode = false;
-        self.mouse_drag_word_origin = None;
-        self.mouse_drag_active = false;
-        // Switch to the group that owns this window.
-        self.focus_group_for_window(window_id);
-        self.set_cursor_for_window(window_id, line, col);
-    }
-
-    /// Handle mouse drag to the given buffer position.
-    /// On first drag: enters Visual mode with anchor at current cursor.
-    /// On subsequent drags: extends selection by moving cursor.
-    /// If already in Visual mode (e.g. from double-click word select),
-    /// preserves the existing anchor and just extends.
-    pub fn mouse_drag(&mut self, window_id: WindowId, line: usize, col: usize) {
-        // Ensure this window's group and tab are active.
-        self.focus_group_for_window(window_id);
-        if self.windows.contains_key(&window_id) {
-            self.active_tab_mut().active_window = window_id;
-        }
-
-        if !self.mouse_drag_active {
-            // First drag event — only set anchor if not already in visual mode
-            // (double-click word select already set the anchor at word start).
-            // Also require the drag to actually reach a *different* buffer position
-            // than the click origin; sub-character mouse jitter on mousedown otherwise
-            // silently enters visual mode before `:` can be pressed.
-            let cursor = self.view().cursor;
-            let moved = line != cursor.line || col != cursor.col;
-            if !moved {
-                return; // Sub-pixel jitter at same cell — ignore.
-            }
-            if !matches!(
-                self.mode,
-                Mode::Visual | Mode::VisualLine | Mode::VisualBlock
-            ) {
-                self.visual_anchor = Some(cursor);
-                self.mode = Mode::Visual;
-            }
-            self.mouse_drag_active = true;
-        }
-
-        // Move cursor to drag position (extends visual selection)
-        let buffer = self.buffer();
-        let max_line = buffer.content.len_lines().saturating_sub(1);
-        let clamped_line = line.min(max_line);
-        let max_col = self.get_max_cursor_col(clamped_line);
-        let clamped_col = col.min(max_col);
-
-        if self.mouse_drag_word_mode {
-            // Word-wise drag: snap to word boundaries
-            if let Some((orig_start, orig_end, orig_line)) = self.mouse_drag_word_origin {
-                let line_text: Vec<char> =
-                    self.buffer().content.line(clamped_line).chars().collect();
-                let drag_before_origin = clamped_line < orig_line
-                    || (clamped_line == orig_line && clamped_col < orig_start);
-
-                if drag_before_origin {
-                    // Dragging before the original word — anchor at word end, cursor at word start
-                    let mut word_start = clamped_col.min(line_text.len().saturating_sub(1));
-                    if word_start < line_text.len() && Self::is_word_char(line_text[word_start]) {
-                        while word_start > 0 && Self::is_word_char(line_text[word_start - 1]) {
-                            word_start -= 1;
-                        }
-                    }
-                    self.visual_anchor = Some(Cursor {
-                        line: orig_line,
-                        col: orig_end,
-                    });
-                    let view = self.view_mut();
-                    view.cursor.line = clamped_line;
-                    view.cursor.col = word_start;
-                } else {
-                    // Dragging after the original word — anchor at word start, cursor at word end
-                    let mut word_end = clamped_col.min(line_text.len().saturating_sub(1));
-                    if word_end < line_text.len() && Self::is_word_char(line_text[word_end]) {
-                        while word_end + 1 < line_text.len()
-                            && Self::is_word_char(line_text[word_end + 1])
-                        {
-                            word_end += 1;
-                        }
-                    }
-                    // Exclude trailing newline
-                    if word_end < line_text.len() && line_text[word_end] == '\n' && word_end > 0 {
-                        word_end -= 1;
-                    }
-                    self.visual_anchor = Some(Cursor {
-                        line: orig_line,
-                        col: orig_start,
-                    });
-                    let view = self.view_mut();
-                    view.cursor.line = clamped_line;
-                    view.cursor.col = word_end;
-                }
-            }
-        } else {
-            let view = self.view_mut();
-            view.cursor.line = clamped_line;
-            view.cursor.col = clamped_col;
-        }
-    }
-
-    /// Handle mouse double-click: select the word under the cursor.
-    /// Positions cursor, finds word boundaries, enters Visual mode.
-    pub fn mouse_double_click(&mut self, window_id: WindowId, line: usize, col: usize) {
-        self.mouse_drag_active = false;
-        self.mouse_drag_word_mode = false;
-        self.mouse_drag_word_origin = None;
-        self.focus_group_for_window(window_id);
-        self.set_cursor_for_window(window_id, line, col);
-
-        // Find word boundaries at cursor
-        let cursor_line = self.view().cursor.line;
-        let cursor_col = self.view().cursor.col;
-        let line_text: Vec<char> = self.buffer().content.line(cursor_line).chars().collect();
-
-        if cursor_col >= line_text.len() || !Self::is_word_char(line_text[cursor_col]) {
-            // Clicked on non-word character — don't select
-            return;
-        }
-
-        // Find word start
-        let mut word_start = cursor_col;
-        while word_start > 0 && Self::is_word_char(line_text[word_start - 1]) {
-            word_start -= 1;
-        }
-
-        // Find word end (inclusive)
-        let mut word_end = cursor_col;
-        while word_end + 1 < line_text.len() && Self::is_word_char(line_text[word_end + 1]) {
-            word_end += 1;
-        }
-        // Exclude trailing newline from word end
-        if word_end < line_text.len() && line_text[word_end] == '\n' && word_end > word_start {
-            word_end -= 1;
-        }
-
-        // Enter visual mode with anchor at word start, cursor at word end
-        self.visual_anchor = Some(Cursor {
-            line: cursor_line,
-            col: word_start,
-        });
-        let view = self.view_mut();
-        view.cursor.col = word_end;
-        self.mode = Mode::Visual;
-        self.mouse_drag_word_mode = true;
-        self.mouse_drag_word_origin = Some((word_start, word_end, cursor_line));
-    }
-
-    // =======================================================================
-    // Clipboard paste into command/search mode
-    // =======================================================================
-
-    /// Paste the first line from the system clipboard into the command buffer.
-    /// Works in Command and Search modes. For Search mode with incremental search,
-    /// also triggers a search update.
-    #[allow(dead_code)]
-    pub fn paste_clipboard_to_input(&mut self) {
-        let text = match self.clipboard_read {
-            Some(ref cb_read) => match cb_read() {
-                Ok(t) => t,
-                Err(e) => {
-                    self.message = format!("Clipboard read failed: {}", e);
-                    return;
-                }
-            },
-            None => return,
-        };
-        self.paste_text_to_input(&text);
-    }
-
-    /// Paste the given text into the command/search buffer (first line only).
-    /// Called by backends that have already fetched the clipboard text themselves.
-    pub fn paste_text_to_input(&mut self, text: &str) {
-        let first_line = text.lines().next().unwrap_or("");
-        if first_line.is_empty() {
-            return;
-        }
-        match self.mode {
-            Mode::Command | Mode::Search => {
-                self.command_insert_str(first_line);
-                if self.mode == Mode::Search && self.settings.incremental_search {
-                    self.perform_incremental_search();
-                }
-            }
-            _ => {}
-        }
-    }
-
-    /// Pre-load clipboard text into the `+` and `*` registers.
-    /// Called by GTK backend after an async GDK clipboard read, before paste.
-    pub fn load_clipboard_register(&mut self, text: String) {
-        self.registers.insert('+', (text.clone(), false));
-        self.registers.insert('*', (text, false));
-    }
-
-    /// Pre-load clipboard text into `"`, `+`, and `*` before a p/P keypress.
-    ///
-    /// If the clipboard content exactly matches what is already in `"`, the
-    /// existing `is_linewise` flag is **preserved** — this covers the common
-    /// `yy` → `p` flow where the yank wrote linewise text to both the register
-    /// and the system clipboard.  When the clipboard holds different text (from
-    /// another application) `is_linewise` is set to `false` as usual.
-    pub fn load_clipboard_for_paste(&mut self, text: String) {
-        let existing_lw = self
-            .registers
-            .get(&'"')
-            .map(|(c, lw)| c == &text && *lw)
-            .unwrap_or(false);
-        self.registers.insert('"', (text.clone(), existing_lw));
-        self.registers.insert('+', (text.clone(), false));
-        self.registers.insert('*', (text, false));
-    }
-
-    // =======================================================================
-    // Toggle case (~)
-    // =======================================================================
-
-    /// Toggle the case of `count` characters starting at the cursor, advance cursor.
-    fn toggle_case_at_cursor(&mut self, count: usize, changed: &mut bool) {
-        let line = self.view().cursor.line;
-        let col = self.view().cursor.col;
-        let char_idx = self.buffer().line_to_char(line) + col;
-
-        // How many chars are available on this line (excluding trailing newline)?
-        let line_len = self.buffer().line_len_chars(line);
-        let line_content = self.buffer().content.line(line);
-        let available = if line_content.chars().last() == Some('\n') {
-            line_len.saturating_sub(1)
-        } else {
-            line_len
-        };
-        let remaining = available.saturating_sub(col);
-        let to_toggle = count.min(remaining);
-
-        if to_toggle == 0 {
-            return;
-        }
-
-        // Read chars to toggle
-        let chars: Vec<char> = self
-            .buffer()
-            .content
-            .slice(char_idx..char_idx + to_toggle)
-            .chars()
-            .collect();
-
-        // Build replacement: toggle case of each char
-        let toggled: String = chars
-            .iter()
-            .map(|&c| {
-                if c.is_uppercase() {
-                    c.to_lowercase().next().unwrap_or(c)
-                } else if c.is_lowercase() {
-                    c.to_uppercase().next().unwrap_or(c)
-                } else {
-                    c
-                }
-            })
-            .collect();
-
-        self.start_undo_group();
-        self.delete_with_undo(char_idx, char_idx + to_toggle);
-        self.insert_with_undo(char_idx, &toggled);
-        self.finish_undo_group();
-
-        // Advance cursor by number of chars toggled (clamped to line end)
-        let new_col = (col + to_toggle).min(available.saturating_sub(1));
-        self.view_mut().cursor.col = new_col;
-        self.clamp_cursor_col();
-        *changed = true;
-    }
-
-    // =======================================================================
-    // Join lines (J)
-    // =======================================================================
-
-    /// Join `count` lines starting at cursor. Collapses the newline + leading
-    /// whitespace of the next line into a single space (no space before `)`).
-    fn join_lines(&mut self, count: usize, changed: &mut bool) {
-        let total_lines = self.buffer().len_lines();
-        let start_line = self.view().cursor.line;
-
-        // We join (count) times; each join merges current line with next
-        let joins = count.min(total_lines.saturating_sub(start_line + 1));
-        if joins == 0 {
-            return;
-        }
-
-        self.start_undo_group();
-        for _ in 0..joins {
-            let cur_line = self.view().cursor.line;
-            let next_line = cur_line + 1;
-            if next_line >= self.buffer().len_lines() {
-                break;
-            }
-
-            // Find position of newline at end of current line
-            let cur_line_len = self.buffer().line_len_chars(cur_line);
-            let cur_line_start = self.buffer().line_to_char(cur_line);
-            // The newline is the last char of the current line
-            let newline_pos = cur_line_start + cur_line_len - 1;
-
-            // Count leading whitespace on next line
-            let next_line_start = self.buffer().line_to_char(next_line);
-            let next_line_content: String = self.buffer().content.line(next_line).chars().collect();
-            let leading_ws = next_line_content
-                .chars()
-                .take_while(|c| *c == ' ' || *c == '\t')
-                .count();
-
-            // Determine what char comes after the whitespace on the next line
-            let next_non_ws = next_line_content.chars().nth(leading_ws);
-
-            // Delete: newline + leading whitespace of next line
-            let del_end = next_line_start + leading_ws;
-            self.delete_with_undo(newline_pos, del_end);
-
-            // Insert a space unless the next non-ws char is ')' or next line was empty/only ws
-            // Also don't add space if the current line ends with a space
-            let should_add_space = !matches!(next_non_ws, None | Some(')') | Some(']') | Some('}'));
-            // Check if current line ends with space (after the newline was removed)
-            let cur_end_char =
-                self.buffer().line_to_char(cur_line) + self.buffer().line_len_chars(cur_line);
-            let ends_with_space = cur_end_char > self.buffer().line_to_char(cur_line)
-                && self.buffer().content.char(cur_end_char - 1) == ' ';
-
-            if should_add_space && !ends_with_space {
-                self.insert_with_undo(newline_pos, " ");
-            }
-        }
-        self.finish_undo_group();
-
-        // Cursor stays at start of original line
-        self.clamp_cursor_col();
-        *changed = true;
-    }
-
-    // =======================================================================
-    // Scroll cursor to position (zz / zt / zb)
-    // =======================================================================
-
-    /// Scroll so that cursor line is centered in viewport.
-    fn scroll_cursor_center(&mut self) {
-        let cursor_line = self.view().cursor.line;
-        let half = self.viewport_lines() / 2;
-        let new_top = cursor_line.saturating_sub(half);
-        self.view_mut().scroll_top = new_top;
-    }
-
-    /// Scroll so that cursor line is at the top of viewport.
-    fn scroll_cursor_top(&mut self) {
-        let cursor_line = self.view().cursor.line;
-        self.view_mut().scroll_top = cursor_line;
-    }
-
-    /// Scroll so that cursor line is at the bottom of viewport.
-    fn scroll_cursor_bottom(&mut self) {
-        let cursor_line = self.view().cursor.line;
-        let viewport = self.viewport_lines();
-        let new_top = cursor_line.saturating_sub(viewport.saturating_sub(1));
-        self.view_mut().scroll_top = new_top;
-    }
-
-    // =======================================================================
-    // Search word under cursor (* / #)
-    // =======================================================================
-
-    /// Extract the word under the cursor. Returns None if cursor is not on a word char.
-    fn word_under_cursor(&self) -> Option<String> {
-        let line = self.view().cursor.line;
-        let col = self.view().cursor.col;
-        let line_content: String = self.buffer().content.line(line).chars().collect();
-        let chars: Vec<char> = line_content.chars().collect();
-
-        if col >= chars.len() {
-            return None;
-        }
-        if !Self::is_word_char(chars[col]) {
-            return None;
-        }
-
-        // Find start of word
-        let start = (0..=col)
-            .rev()
-            .take_while(|&i| Self::is_word_char(chars[i]))
-            .last()
-            .unwrap_or(col);
-        // Find end of word (exclusive)
-        let end = (col..chars.len())
-            .take_while(|&i| Self::is_word_char(chars[i]))
-            .last()
-            .map(|i| i + 1)
-            .unwrap_or(col + 1);
-
-        Some(chars[start..end].iter().collect())
-    }
-
-    /// Search forward (*) or backward (#) for the word under cursor with word boundaries.
-    fn search_word_under_cursor(&mut self, forward: bool) {
-        let word = match self.word_under_cursor() {
-            Some(w) => w,
-            None => {
-                self.message = "No word under cursor".to_string();
-                return;
-            }
-        };
-
-        self.search_query = word.clone();
-        self.search_direction = if forward {
-            SearchDirection::Forward
-        } else {
-            SearchDirection::Backward
-        };
-        self.search_word_bounded = true;
-
-        // Build word-boundary matches manually
-        self.build_word_bounded_matches();
-
-        if self.search_matches.is_empty() {
-            self.message = format!("Pattern not found: {}", word);
-            return;
-        }
-
-        // Jump to first match in the appropriate direction
-        if forward {
-            self.search_next();
-        } else {
-            self.search_prev();
-        }
-    }
-
-    /// Like run_search but only keeps matches that are whole words.
-    fn build_word_bounded_matches(&mut self) {
-        self.search_matches.clear();
-        self.search_index = None;
-
-        if self.search_query.is_empty() {
-            return;
-        }
-
-        let text = self.buffer().to_string();
-        let query = self.search_query.clone();
-        let mut byte_pos = 0;
-
-        while let Some(found) = text[byte_pos..].find(&query) {
-            let start_byte = byte_pos + found;
-            let end_byte = start_byte + query.len();
-
-            // Check word boundaries
-            let before_ok = start_byte == 0 || {
-                let c = text[..start_byte].chars().last().unwrap_or(' ');
-                !Self::is_word_char(c)
-            };
-            let after_ok = end_byte >= text.len() || {
-                let c = text[end_byte..].chars().next().unwrap_or(' ');
-                !Self::is_word_char(c)
-            };
-
-            if before_ok && after_ok {
-                let start_char = self.buffer().content.byte_to_char(start_byte);
-                let end_char = self.buffer().content.byte_to_char(end_byte);
-                self.search_matches.push((start_char, end_char));
-            }
-
-            byte_pos = start_byte + 1;
-        }
-    }
-
-    // =======================================================================
-    // Multiple cursors (Alt-D)
-    // =======================================================================
-
-    /// Convert a char index in the current buffer into a `Cursor` (line, col).
-    fn char_idx_to_cursor(&self, char_idx: usize) -> Cursor {
-        let len = self.buffer().content.len_chars();
-        let char_idx = char_idx.min(len);
-        if len == 0 {
-            return Cursor { line: 0, col: 0 };
-        }
-        let line = self.buffer().content.char_to_line(char_idx);
-        let line_start = self.buffer().line_to_char(line);
-        Cursor {
-            line,
-            col: char_idx - line_start,
-        }
-    }
-
-    /// Convert a byte offset in the buffer text into a `Cursor`.
-    fn byte_offset_to_cursor(&self, byte_offset: usize) -> Cursor {
-        let char_idx = self.buffer().content.byte_to_char(byte_offset);
-        self.char_idx_to_cursor(char_idx)
-    }
-
-    /// Search for the next occurrence of `pattern` in the buffer, starting
-    /// one pattern-length past `after`.  Wraps around the document end.
-    /// Returns `None` if `pattern` is not found anywhere in the buffer.
-    fn find_next_occurrence(
-        &self,
-        pattern: &str,
-        after: Cursor,
-        word_bounded: bool,
-    ) -> Option<Cursor> {
-        if pattern.is_empty() {
-            return None;
-        }
-
-        let text = self.buffer().to_string();
-        // Start searching one pattern-length past the given cursor position.
-        let after_char_idx =
-            self.buffer().line_to_char(after.line) + after.col + pattern.chars().count();
-        let after_byte = self
-            .buffer()
-            .content
-            .char_to_byte(after_char_idx.min(self.buffer().content.len_chars()));
-
-        let check_boundary = |sb: usize, eb: usize| -> bool {
-            if !word_bounded {
-                return true;
-            }
-            let before_ok =
-                sb == 0 || !Self::is_word_char(text[..sb].chars().last().unwrap_or(' '));
-            let after_ok =
-                eb >= text.len() || !Self::is_word_char(text[eb..].chars().next().unwrap_or(' '));
-            before_ok && after_ok
-        };
-
-        // Pass 1: from after_byte to end of document.
-        let mut byte_pos = after_byte;
-        while byte_pos < text.len() {
-            match text[byte_pos..].find(pattern) {
-                None => break,
-                Some(found) => {
-                    let sb = byte_pos + found;
-                    let eb = sb + pattern.len();
-                    if check_boundary(sb, eb) {
-                        return Some(self.byte_offset_to_cursor(sb));
-                    }
-                    byte_pos = sb + 1;
-                }
-            }
-        }
-
-        // Pass 2: wrap around from document start to after_byte.
-        byte_pos = 0;
-        while byte_pos < after_byte {
-            match text[byte_pos..].find(pattern) {
-                None => break,
-                Some(found) => {
-                    let sb = byte_pos + found;
-                    if sb >= after_byte {
-                        break;
-                    }
-                    let eb = sb + pattern.len();
-                    if check_boundary(sb, eb) {
-                        return Some(self.byte_offset_to_cursor(sb));
-                    }
-                    byte_pos = sb + 1;
-                }
-            }
-        }
-
-        None
-    }
-
-    /// Collect all byte-offset positions of `pattern` in the current buffer,
-    /// returning them as `Cursor` values.  When `word_bounded` is true only
-    /// whole-word matches are returned.
-    fn collect_all_occurrences(&self, pattern: &str, word_bounded: bool) -> Vec<Cursor> {
-        if pattern.is_empty() {
-            return vec![];
-        }
-        let text = self.buffer().to_string();
-        let mut results = Vec::new();
-        let mut byte_pos = 0;
-        while byte_pos < text.len() {
-            match text[byte_pos..].find(pattern) {
-                None => break,
-                Some(found) => {
-                    let sb = byte_pos + found;
-                    let eb = sb + pattern.len();
-                    let ok = if word_bounded {
-                        let before_ok = sb == 0
-                            || !Self::is_word_char(text[..sb].chars().last().unwrap_or(' '));
-                        let after_ok = eb >= text.len()
-                            || !Self::is_word_char(text[eb..].chars().next().unwrap_or(' '));
-                        before_ok && after_ok
-                    } else {
-                        true
-                    };
-                    if ok {
-                        results.push(self.byte_offset_to_cursor(sb));
-                    }
-                    byte_pos = sb + 1;
-                }
-            }
-        }
-        results
-    }
-
-    /// Add secondary cursors at *every* occurrence of the word under the
-    /// primary cursor.  Called by backends when `select_all_matches` is pressed.
-    pub fn select_all_word_occurrences(&mut self) -> EngineAction {
-        let word = match self.word_under_cursor() {
-            Some(w) => w,
-            None => {
-                self.message = "No word under cursor".to_string();
-                return EngineAction::None;
-            }
-        };
-        let all = self.collect_all_occurrences(&word, true);
-        if all.is_empty() {
-            self.message = format!("No occurrences of '{}'", word);
-            return EngineAction::None;
-        }
-        let primary = *self.cursor();
-        let extras: Vec<Cursor> = all.into_iter().filter(|&c| c != primary).collect();
-        let n = extras.len();
-        self.view_mut().extra_cursors = extras;
-        self.message = format!("{} cursors (all occurrences of '{}')", n + 1, word);
-        EngineAction::None
-    }
-
-    /// Add a secondary cursor at the given `(line, col)` position.
-    /// Does nothing if the position equals the primary cursor or is already
-    /// present in `extra_cursors`.
-    pub fn add_cursor_at_pos(&mut self, line: usize, col: usize) {
-        let new_cursor = Cursor { line, col };
-        if new_cursor == *self.cursor() {
-            return;
-        }
-        if self.view().extra_cursors.contains(&new_cursor) {
-            return;
-        }
-        self.view_mut().extra_cursors.push(new_cursor);
-    }
-
-    /// Add a secondary cursor at the next occurrence of the word under the
-    /// primary cursor (or after the last extra cursor if any exist).
-    /// Called by backends when the configured `add_cursor` key is pressed.
-    pub fn add_cursor_at_next_match(&mut self) -> EngineAction {
-        let word = match self.word_under_cursor() {
-            Some(w) => w,
-            None => {
-                self.message = "No word under cursor".to_string();
-                return EngineAction::None;
-            }
-        };
-        let search_after = self
-            .view()
-            .extra_cursors
-            .last()
-            .copied()
-            .unwrap_or_else(|| *self.cursor());
-        if let Some(new_cursor) = self.find_next_occurrence(&word, search_after, true) {
-            let is_primary = new_cursor == *self.cursor();
-            let already_extra = self.view().extra_cursors.contains(&new_cursor);
-            if !is_primary && !already_extra {
-                self.view_mut().extra_cursors.push(new_cursor);
-                let total = self.view().extra_cursors.len() + 1; // +1 for primary
-                self.message = format!("{} cursors ('{}')", total, word);
-            } else {
-                self.message = format!("No more occurrences of '{}'", word);
-            }
-        } else {
-            self.message = format!("No more occurrences of '{}'", word);
-        }
-        EngineAction::None
-    }
-
-    // ── Multi-cursor editing helpers ─────────────────────────────────────────
-
-    /// Insert `text` at every cursor position (primary + extra) simultaneously.
-    /// Processes in ascending char-index order with a running offset so that
-    /// each subsequent insert uses the correct adjusted position.
-    /// Updates primary cursor and all extra cursors to point just after their
-    /// respective inserted text.
-    fn mc_insert(&mut self, text: &str) {
-        let extra = self.view().extra_cursors.clone();
-        let primary = *self.cursor();
-
-        let primary_orig = self.buffer().line_to_char(primary.line) + primary.col;
-        let extra_origs: Vec<usize> = extra
-            .iter()
-            .map(|c| self.buffer().line_to_char(c.line) + c.col)
-            .collect();
-
-        // All original char indices sorted ascending (safe to sort since positions are distinct).
-        let mut all_origs: Vec<usize> = extra_origs.clone();
-        all_origs.push(primary_orig);
-        all_origs.sort_unstable();
-
-        let insert_chars = text.chars().count();
-
-        // Pre-compute new char indices before modifying the buffer.
-        // Cursor at ascending rank i → new_cidx = orig + (rank+1)*insert_chars.
-        let rank_of = |orig: usize| all_origs.iter().position(|&x| x == orig).unwrap_or(0);
-        let primary_new_cidx = primary_orig + (rank_of(primary_orig) + 1) * insert_chars;
-        let extra_new_cidxs: Vec<usize> = extra_origs
-            .iter()
-            .map(|&orig| orig + (rank_of(orig) + 1) * insert_chars)
-            .collect();
-
-        // Insert in ascending order with cumulative offset.
-        let mut offset = 0usize;
-        for &orig in &all_origs {
-            self.insert_with_undo(orig + offset, text);
-            offset += insert_chars;
-        }
-
-        // Apply updated positions (buffer is now modified; char_idx_to_cursor uses new state).
-        self.view_mut().cursor = self.char_idx_to_cursor(primary_new_cidx);
-        self.view_mut().extra_cursors = extra_new_cidxs
-            .iter()
-            .map(|&cidx| self.char_idx_to_cursor(cidx))
-            .collect();
-    }
-
-    /// Delete one char before every cursor position with col > 0.
-    /// Extra cursors at col == 0 are left in place (line-merge not done in multi-cursor mode).
-    /// Returns `true` if at least one deletion was performed.
-    fn mc_backspace(&mut self) -> bool {
-        let extra = self.view().extra_cursors.clone();
-        let primary = *self.cursor();
-
-        let primary_orig = self.buffer().line_to_char(primary.line) + primary.col;
-
-        // Pre-compute original char indices for extra cursors (before any modification).
-        let extra_data: Vec<(usize, bool)> = extra
-            .iter()
-            .map(|c| {
-                let orig = self.buffer().line_to_char(c.line) + c.col;
-                (orig, c.col > 0)
-            })
-            .collect();
-
-        // Collect eligible (col > 0) original char indices.
-        let mut all_eligible: Vec<usize> = Vec::new();
-        let primary_eligible = primary.col > 0;
-        if primary_eligible {
-            all_eligible.push(primary_orig);
-        }
-        for &(orig, eligible) in &extra_data {
-            if eligible {
-                all_eligible.push(orig);
-            }
-        }
-
-        if all_eligible.is_empty() {
-            return false;
-        }
-
-        all_eligible.sort_unstable();
-
-        // Pre-compute new char indices before modifying the buffer.
-        // Cursor at ascending rank i → new_cidx = orig - (rank+1).
-        let rank_of = |orig: usize| all_eligible.iter().position(|&x| x == orig).unwrap_or(0);
-        let primary_new_cidx = if primary_eligible {
-            Some(primary_orig - (rank_of(primary_orig) + 1))
-        } else {
-            None
-        };
-        let extra_new_cidxs: Vec<Option<usize>> = extra_data
-            .iter()
-            .map(|&(orig, eligible)| {
-                if eligible {
-                    Some(orig - (rank_of(orig) + 1))
-                } else {
-                    None
-                }
-            })
-            .collect();
-
-        // Delete in DESCENDING order (no offset adjustment needed).
-        for &orig in all_eligible.iter().rev() {
-            self.delete_with_undo(orig - 1, orig);
-        }
-
-        // Apply updated positions.
-        if let Some(new_cidx) = primary_new_cidx {
-            self.view_mut().cursor = self.char_idx_to_cursor(new_cidx);
-        }
-        self.view_mut().extra_cursors = extra_new_cidxs
-            .iter()
-            .zip(extra.iter())
-            .map(|(&opt, ec)| {
-                if let Some(new_cidx) = opt {
-                    self.char_idx_to_cursor(new_cidx)
-                } else {
-                    *ec // unchanged (was at col == 0)
-                }
-            })
-            .collect();
-
-        true
-    }
-
-    /// Delete one char after every cursor position that is not at end-of-buffer.
-    /// Returns `true` if at least one deletion was performed.
-    fn mc_delete_forward(&mut self) -> bool {
-        let extra = self.view().extra_cursors.clone();
-        let primary = *self.cursor();
-
-        let buf_len = self.buffer().content.len_chars();
-        let primary_orig = self.buffer().line_to_char(primary.line) + primary.col;
-
-        let extra_data: Vec<(usize, bool)> = extra
-            .iter()
-            .map(|c| {
-                let orig = self.buffer().line_to_char(c.line) + c.col;
-                (orig, orig < buf_len)
-            })
-            .collect();
-
-        let mut all_eligible: Vec<usize> = Vec::new();
-        let primary_eligible = primary_orig < buf_len;
-        if primary_eligible {
-            all_eligible.push(primary_orig);
-        }
-        for &(orig, eligible) in &extra_data {
-            if eligible {
-                all_eligible.push(orig);
-            }
-        }
-
-        if all_eligible.is_empty() {
-            return false;
-        }
-
-        all_eligible.sort_unstable();
-
-        // Pre-compute new char indices.
-        // Delete-forward: cursor stays in place; earlier deletions shift it left.
-        // Cursor at ascending rank i → new_cidx = orig - rank (not rank+1).
-        let rank_of = |orig: usize| all_eligible.iter().position(|&x| x == orig).unwrap_or(0);
-        let primary_new_cidx = if primary_eligible {
-            Some(primary_orig - rank_of(primary_orig))
-        } else {
-            None
-        };
-        let extra_new_cidxs: Vec<Option<usize>> = extra_data
-            .iter()
-            .map(|&(orig, eligible)| {
-                if eligible {
-                    Some(orig - rank_of(orig))
-                } else {
-                    None
-                }
-            })
-            .collect();
-
-        // Delete in DESCENDING order.
-        for &orig in all_eligible.iter().rev() {
-            self.delete_with_undo(orig, orig + 1);
-        }
-
-        if let Some(new_cidx) = primary_new_cidx {
-            self.view_mut().cursor = self.char_idx_to_cursor(new_cidx);
-        }
-        self.view_mut().extra_cursors = extra_new_cidxs
-            .iter()
-            .zip(extra.iter())
-            .map(|(&opt, ec)| {
-                if let Some(new_cidx) = opt {
-                    self.char_idx_to_cursor(new_cidx)
-                } else {
-                    *ec
-                }
-            })
-            .collect();
-
-        true
-    }
-
-    /// Insert a newline (+ auto-indent) at every cursor position.
-    /// Each cursor gets the indent of its own line computed before any modification.
-    fn mc_return(&mut self) {
-        let extra = self.view().extra_cursors.clone();
-        let primary = *self.cursor();
-
-        // Pre-compute (orig_cidx, insert_text) for every cursor, ascending.
-        struct ReturnOp {
-            orig_cidx: usize,
-            text: String,
-            is_primary: bool,
-            extra_idx: usize,
-        }
-
-        let primary_indent = if self.settings.auto_indent {
-            self.get_line_indent_str(primary.line)
-        } else {
-            String::new()
-        };
-
-        let extra_ops: Vec<(usize, String)> = extra
-            .iter()
-            .map(|c| {
-                let orig = self.buffer().line_to_char(c.line) + c.col;
-                let indent = if self.settings.auto_indent {
-                    self.get_line_indent_str(c.line)
-                } else {
-                    String::new()
-                };
-                (orig, format!("\n{}", indent))
-            })
-            .collect();
-
-        let primary_orig = self.buffer().line_to_char(primary.line) + primary.col;
-        let primary_text = format!("\n{}", primary_indent);
-
-        let mut all_ops: Vec<ReturnOp> = extra_ops
-            .iter()
-            .enumerate()
-            .map(|(i, (orig, text))| ReturnOp {
-                orig_cidx: *orig,
-                text: text.clone(),
-                is_primary: false,
-                extra_idx: i,
-            })
-            .collect();
-        all_ops.push(ReturnOp {
-            orig_cidx: primary_orig,
-            text: primary_text,
-            is_primary: true,
-            extra_idx: 0,
-        });
-        all_ops.sort_by_key(|op| op.orig_cidx);
-
-        // Apply inserts ascending with cumulative offset; cursor goes to end of each insert.
-        let mut running_offset = 0usize;
-        let mut primary_new_cidx = 0usize;
-        let mut extra_new_cidxs = vec![0usize; extra.len()];
-
-        for op in &all_ops {
-            let text_chars = op.text.chars().count();
-            let insert_at = op.orig_cidx + running_offset;
-            self.insert_with_undo(insert_at, &op.text);
-            let new_cidx = insert_at + text_chars;
-            running_offset += text_chars;
-            if op.is_primary {
-                primary_new_cidx = new_cidx;
-            } else {
-                extra_new_cidxs[op.extra_idx] = new_cidx;
-            }
-        }
-
-        self.view_mut().cursor = self.char_idx_to_cursor(primary_new_cidx);
-        self.view_mut().extra_cursors = extra_new_cidxs
-            .iter()
-            .map(|&cidx| self.char_idx_to_cursor(cidx))
-            .collect();
-    }
-
-    // =======================================================================
-    // Jump list (Ctrl-O / Ctrl-I)
-    // =======================================================================
-
-    /// Push (line, col) to the change list, capped at 100 entries.
-    fn push_change_location(&mut self, line: usize, col: usize) {
-        // Truncate any forward entries (if we navigated back with g;)
-        self.change_list.truncate(self.change_list_pos);
-        // Avoid duplicate consecutive entries
-        if self.change_list.last() == Some(&(line, col)) {
-            return;
-        }
-        self.change_list.push((line, col));
-        if self.change_list.len() > 100 {
-            self.change_list.remove(0);
-        }
-        self.change_list_pos = self.change_list.len();
-    }
-
-    /// Push the current cursor position onto the jump list.
-    pub fn push_jump_location(&mut self) {
-        // Save pre-jump position for '' / `` marks
-        let line = self.view().cursor.line;
-        let col = self.view().cursor.col;
-        self.last_jump_pos = Some((line, col));
-
-        let file = self.active_buffer_state().file_path.clone();
-        let line = self.view().cursor.line;
-        let col = self.view().cursor.col;
-
-        // Truncate forward history when a new jump is made
-        if self.jump_list_pos < self.jump_list.len() {
-            self.jump_list.truncate(self.jump_list_pos);
-        }
-
-        // Don't push a duplicate of the current top entry
-        if let Some(last) = self.jump_list.last() {
-            if last.0 == file && last.1 == line && last.2 == col {
-                return;
-            }
-        }
-
-        self.jump_list.push((file, line, col));
-
-        // Cap at 100 entries
-        if self.jump_list.len() > 100 {
-            self.jump_list.remove(0);
-        }
-
-        self.jump_list_pos = self.jump_list.len();
-    }
-
-    /// Navigate backward in the jump list (Ctrl-O).
-    pub fn jump_list_back(&mut self) {
-        // When at the "live" end (not stored in list), save current position
-        // so Ctrl-I can return to it, then jump to the previous entry.
-        if self.jump_list_pos == self.jump_list.len() {
-            if self.jump_list.is_empty() {
-                self.message = "Already at oldest position in jump list".to_string();
-                return;
-            }
-            let file = self.active_buffer_state().file_path.clone();
-            let line = self.view().cursor.line;
-            let col = self.view().cursor.col;
-            #[allow(clippy::unnecessary_map_or)] // is_none_or requires Rust 1.82+
-            let should_push = self.jump_list.last().map_or(true, |last| {
-                last.0 != file || last.1 != line || last.2 != col
-            });
-            if should_push {
-                self.jump_list.push((file, line, col));
-                if self.jump_list.len() > 100 {
-                    self.jump_list.remove(0);
-                }
-            }
-            // Jump to the entry BEFORE the one we just saved
-            // (list.len()-1 is current, list.len()-2 is the previous)
-            if self.jump_list.len() < 2 {
-                self.message = "Already at oldest position in jump list".to_string();
-                return;
-            }
-            self.jump_list_pos = self.jump_list.len() - 2;
-            self.apply_jump_list_entry(self.jump_list_pos);
-            return;
-        }
-
-        // We're inside the list — go to the previous entry
-        if self.jump_list_pos == 0 {
-            self.message = "Already at oldest position in jump list".to_string();
-            return;
-        }
-
-        self.jump_list_pos -= 1;
-        self.apply_jump_list_entry(self.jump_list_pos);
-    }
-
-    /// Navigate forward in the jump list (Ctrl-I / Tab).
-    pub fn jump_list_forward(&mut self) {
-        if self.jump_list_pos + 1 >= self.jump_list.len() {
-            self.message = "Already at newest position in jump list".to_string();
-            return;
-        }
-
-        self.jump_list_pos += 1;
-        self.apply_jump_list_entry(self.jump_list_pos);
-    }
-
-    /// Move to the position stored at the given jump list index.
-    fn apply_jump_list_entry(&mut self, idx: usize) {
-        let entry = match self.jump_list.get(idx) {
-            Some(e) => e.clone(),
-            None => return,
-        };
-
-        let (file, line, col) = entry;
-
-        // If cross-file, open the file
-        let current_file = self.active_buffer_state().file_path.clone();
-        if file != current_file {
-            if let Some(path) = &file {
-                let path = path.clone();
-                let _ = self.open_file_with_mode(&path, OpenMode::Permanent);
-            }
-        }
-
-        let max_line = self.buffer().len_lines().saturating_sub(1);
-        self.view_mut().cursor.line = line.min(max_line);
-        self.view_mut().cursor.col = col;
-        self.clamp_cursor_col();
-    }
-
-    // =======================================================================
-    // Indent / Dedent (>> / <<)
-    // =======================================================================
-
-    /// Indent `count` lines starting at `start_line` by shift_width.
-    fn indent_lines(&mut self, start_line: usize, count: usize, changed: &mut bool) {
-        let indent_str = if self.settings.expand_tab {
-            " ".repeat(self.settings.shift_width as usize)
-        } else {
-            "\t".to_string()
-        };
-
-        self.start_undo_group();
-        let total = self.buffer().len_lines();
-        for i in 0..count {
-            let line_idx = start_line + i;
-            if line_idx >= total {
-                break;
-            }
-            let line_start = self.buffer().line_to_char(line_idx);
-            self.insert_with_undo(line_start, &indent_str);
-        }
-        self.finish_undo_group();
-        *changed = true;
-    }
-
-    /// Dedent `count` lines starting at `start_line` by up to shift_width.
-    fn dedent_lines(&mut self, start_line: usize, count: usize, changed: &mut bool) {
-        let sw = self.settings.shift_width as usize;
-        self.start_undo_group();
-        // Work backwards to avoid invalidating positions
-        let total = self.buffer().len_lines();
-        for i in (0..count).rev() {
-            let line_idx = start_line + i;
-            if line_idx >= total {
-                continue;
-            }
-            let line_start = self.buffer().line_to_char(line_idx);
-            let line_content: String = self.buffer().content.line(line_idx).chars().collect();
-            let mut removed = 0;
-            for ch in line_content.chars() {
-                if removed >= sw {
-                    break;
-                }
-                match ch {
-                    ' ' => removed += 1,
-                    '\t' => removed += sw.min(sw - (removed % sw).max(1) + 1).min(sw - removed),
-                    _ => break,
-                }
-            }
-            if removed > 0 {
-                self.delete_with_undo(line_start, line_start + removed);
-            }
-        }
-        self.finish_undo_group();
-        if count > 0 {
-            *changed = true;
-        }
     }
 }
